Struct aws_sdk_ecs::client::fluent_builders::CreateService

pub struct CreateService<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateService<C>

pub async fn send(
self
) -> Result<CreateServiceOutput, SdkError<CreateServiceError>> where
C: SmithyConnector,

pub fn cluster(self, input: impl Into<String>) -> Self

The short name or full Amazon Resource Name (ARN) of the cluster on which to run your service. If you do not specify a cluster, the default cluster is assumed.

pub fn set_cluster(self, input: Option<String>) -> Self

pub fn service_name(self, input: impl Into<String>) -> Self

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

pub fn set_service_name(self, input: Option<String>) -> Self

pub fn task_definition(self, input: impl Into<String>) -> Self

The family and revision (family:revision) or full ARN of the task definition to run in your service. If a revision is not specified, the latest ACTIVE revision is used.

A task definition must be specified if the service is using either the ECS or CODE_DEPLOY deployment controllers.

pub fn set_task_definition(self, input: Option<String>) -> Self

pub fn load_balancers(self, inp: impl Into<LoadBalancer>) -> Self

A load balancer object representing the load balancers to use with your service. For more information, see Service Load Balancing in the Amazon Elastic Container Service Developer Guide.

If the service is using the rolling update (ECS) deployment controller and using either an Application Load Balancer or Network Load Balancer, you must specify one or more target group ARNs to attach to the service. The service-linked role is required for services that make use of multiple target groups. For more information, see Using service-linked roles for Amazon ECS in the Amazon Elastic Container Service Developer Guide.

If the service is using the CODE_DEPLOY deployment controller, the service is required to use either an Application Load Balancer or Network Load Balancer. When creating an AWS CodeDeploy deployment group, you specify two target groups (referred to as a targetGroupPair). During a deployment, AWS CodeDeploy determines which task set in your service has the status PRIMARY and associates one target group with it, and then associates the other target group with the replacement task set. The load balancer can also have up to two listeners: a required listener for production traffic and an optional listener that allows you perform validation tests with Lambda functions before routing production traffic to it.

After you create a service using the ECS deployment controller, the load balancer name or target group ARN, container name, and container port specified in the service definition are immutable. If you are using the CODE_DEPLOY deployment controller, these values can be changed when updating the service.

For Application Load Balancers and Network Load Balancers, this object must contain the load balancer target group ARN, the container name (as it appears in a container definition), and the container port to access from the load balancer. The load balancer name parameter must be omitted. When a task from this service is placed on a container instance, the container instance and port combination is registered as a target in the target group specified here.

For Classic Load Balancers, this object must contain the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer. The target group ARN parameter must be omitted. When a task from this service is placed on a container instance, the container instance is registered with the load balancer specified here.

Services with tasks that use the awsvpc network mode (for example, those with the Fargate launch type) only support Application Load Balancers and Network Load Balancers. Classic Load Balancers are not supported. Also, when you create any target groups for these services, you must choose ip as the target type, not instance, because tasks that use the awsvpc network mode are associated with an elastic network interface, not an Amazon EC2 instance.

pub fn set_load_balancers(self, input: Option<Vec<LoadBalancer>>) -> Self

pub fn service_registries(self, inp: impl Into<ServiceRegistry>) -> Self

The details of the service discovery registry to associate with this service. For more information, see Service discovery.

Each service may be associated with one service registry. Multiple service registries per service isn't supported.

pub fn set_service_registries(self, input: Option<Vec<ServiceRegistry>>) -> Self

pub fn desired_count(self, input: i32) -> Self

The number of instantiations of the specified task definition to place and keep running on your cluster.

This is required if schedulingStrategy is REPLICA or is not specified. If schedulingStrategy is DAEMON then this is not required.

pub fn set_desired_count(self, input: Option<i32>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn launch_type(self, input: LaunchType) -> Self

The infrastructure on which to run your service. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

The FARGATE launch type runs your tasks on AWS Fargate On-Demand infrastructure.

Fargate Spot infrastructure is available for use but a capacity provider strategy must be used. For more information, see AWS Fargate capacity providers in the Amazon ECS User Guide for AWS Fargate.

The EC2 launch type runs your tasks on Amazon EC2 instances registered to your cluster.

The EXTERNAL launch type runs your tasks on your on-premise server or virtual machine (VM) capacity registered to your cluster.

A service can use either a launch type or a capacity provider strategy. If a launchType is specified, the capacityProviderStrategy parameter must be omitted.

pub fn set_launch_type(self, input: Option<LaunchType>) -> Self

pub fn capacity_provider_strategy(
self,
inp: impl Into<CapacityProviderStrategyItem>
) -> Self

The capacity provider strategy to use for the service.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

pub fn set_capacity_provider_strategy(
self,
input: Option<Vec<CapacityProviderStrategyItem>>
) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

The platform version that your tasks in the service are running on. A platform version is specified only for tasks using the Fargate launch type. If one isn't specified, the LATEST platform version is used by default. For more information, see AWS Fargate platform versions in the Amazon Elastic Container Service Developer Guide.

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn role(self, input: impl Into<String>) -> Self

The name or full Amazon Resource Name (ARN) of the IAM role that allows Amazon ECS to make calls to your load balancer on your behalf. This parameter is only permitted if you are using a load balancer with your service and your task definition does not use the awsvpc network mode. If you specify the role parameter, you must also specify a load balancer object with the loadBalancers parameter.

If your account has already created the Amazon ECS service-linked role, that role is used by default for your service unless you specify a role here. The service-linked role is required if your task definition uses the awsvpc network mode or if the service is configured to use service discovery, an external deployment controller, multiple target groups, or Elastic Inference accelerators in which case you should not specify a role here. For more information, see Using service-linked roles for Amazon ECS in the Amazon Elastic Container Service Developer Guide.

If your specified role has a path other than /, then you must either specify the full role ARN (this is recommended) or prefix the role name with the path. For example, if a role with the name bar has a path of /foo/ then you would specify /foo/bar as the role name. For more information, see Friendly names and paths in the IAM User Guide.

pub fn set_role(self, input: Option<String>) -> Self

pub fn deployment_configuration(self, input: DeploymentConfiguration) -> Self

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

pub fn set_deployment_configuration(
self,
input: Option<DeploymentConfiguration>
) -> Self

pub fn placement_constraints(self, inp: impl Into<PlacementConstraint>) -> Self

An array of placement constraint objects to use for tasks in your service. You can specify a maximum of 10 constraints per task (this limit includes constraints in the task definition and those specified at runtime).

pub fn set_placement_constraints(
self,
input: Option<Vec<PlacementConstraint>>
) -> Self

pub fn placement_strategy(self, inp: impl Into<PlacementStrategy>) -> Self

The placement strategy objects to use for tasks in your service. You can specify a maximum of five strategy rules per service.

pub fn set_placement_strategy(
self,
input: Option<Vec<PlacementStrategy>>
) -> Self

pub fn network_configuration(self, input: NetworkConfiguration) -> Self

The network configuration for the service. This parameter is required for task definitions that use the awsvpc network mode to receive their own elastic network interface, and it is not supported for other network modes. For more information, see Task networking in the Amazon Elastic Container Service Developer Guide.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

pub fn health_check_grace_period_seconds(self, input: i32) -> Self

The period of time, in seconds, that the Amazon ECS service scheduler should ignore unhealthy Elastic Load Balancing target health checks after a task has first started. This is only used when your service is configured to use a load balancer. If your service has a load balancer defined and you don't specify a health check grace period value, the default value of 0 is used.

If your service's tasks take a while to start and respond to Elastic Load Balancing health checks, you can specify a health check grace period of up to 2,147,483,647 seconds. During that time, the Amazon ECS service scheduler ignores health check status. This grace period can prevent the service scheduler from marking tasks as unhealthy and stopping them before they have time to come up.

pub fn set_health_check_grace_period_seconds(self, input: Option<i32>) -> Self

pub fn scheduling_strategy(self, input: SchedulingStrategy) -> Self

The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions. This scheduler strategy is required if the service is using the CODE_DEPLOY or EXTERNAL deployment controller types.

DAEMON-The daemon scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that you specify in your cluster. The service scheduler also evaluates the task placement constraints for running tasks and will stop tasks that do not meet the placement constraints. When you're using this strategy, you don't need to specify a desired number of tasks, a task placement strategy, or use Service Auto Scaling policies.

Tasks using the Fargate launch type or the CODE_DEPLOY or EXTERNAL deployment controller types don't support the DAEMON scheduling strategy.

pub fn set_scheduling_strategy(self, input: Option<SchedulingStrategy>) -> Self

pub fn deployment_controller(self, input: DeploymentController) -> Self

The deployment controller to use for the service. If no deployment controller is specified, the default value of ECS is used.

pub fn set_deployment_controller(
self,
input: Option<DeploymentController>
) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. When a service is deleted, the tags are deleted as well.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for AWS use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn enable_ecs_managed_tags(self, input: bool) -> Self

Specifies whether to enable Amazon ECS managed tags for the tasks within the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

pub fn set_enable_ecs_managed_tags(self, input: Option<bool>) -> Self

pub fn propagate_tags(self, input: PropagateTags) -> Self

Specifies whether to propagate the tags from the task definition or the service to the tasks in the service. If no value is specified, the tags are not propagated. Tags can only be propagated to the tasks within the service during service creation. To add tags to a task after service creation, use the TagResource API action.

pub fn set_propagate_tags(self, input: Option<PropagateTags>) -> Self

pub fn enable_execute_command(self, input: bool) -> Self

Whether or not the execute command functionality is enabled for the service. If true, this enables execute command functionality on all containers in the service tasks.

pub fn set_enable_execute_command(self, input: Option<bool>) -> Self