Struct aws_sdk_ecs::client::fluent_builders::CreateTaskSet

pub struct CreateTaskSet<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateTaskSet<C>

pub async fn send(
self
) -> Result<CreateTaskSetOutput, SdkError<CreateTaskSetError>> where
C: SmithyConnector,

pub fn service(self, input: impl Into<String>) -> Self

The short name or full Amazon Resource Name (ARN) of the service to create the task set in.

pub fn set_service(self, input: Option<String>) -> Self

pub fn cluster(self, input: impl Into<String>) -> Self

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the service to create the task set in.

pub fn set_cluster(self, input: Option<String>) -> Self

pub fn external_id(self, input: impl Into<String>) -> Self

An optional non-unique tag that identifies this task set in external systems. If the task set is associated with a service discovery registry, the tasks in this task set will have the ECS_TASK_SET_EXTERNAL_ID AWS Cloud Map attribute set to the provided value.

pub fn set_external_id(self, input: Option<String>) -> Self

pub fn task_definition(self, input: impl Into<String>) -> Self

The task definition for the tasks in the task set to use.

pub fn set_task_definition(self, input: Option<String>) -> Self

pub fn network_configuration(self, input: NetworkConfiguration) -> Self

An object representing the network configuration for a task set.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

pub fn load_balancers(self, inp: impl Into<LoadBalancer>) -> Self

A load balancer object representing the load balancer to use with the task set. The supported load balancer types are either an Application Load Balancer or a Network Load Balancer.

pub fn set_load_balancers(self, input: Option<Vec<LoadBalancer>>) -> Self

pub fn service_registries(self, inp: impl Into<ServiceRegistry>) -> Self

The details of the service discovery registries to assign to this task set. For more information, see Service Discovery.

pub fn set_service_registries(self, input: Option<Vec<ServiceRegistry>>) -> Self

pub fn launch_type(self, input: LaunchType) -> Self

The launch type that new tasks in the task set will use. For more information, see Amazon ECS Launch Types in the Amazon Elastic Container Service Developer Guide.

If a launchType is specified, the capacityProviderStrategy parameter must be omitted.

pub fn set_launch_type(self, input: Option<LaunchType>) -> Self

pub fn capacity_provider_strategy(
self,
inp: impl Into<CapacityProviderStrategyItem>
) -> Self

The capacity provider strategy to use for the task set.

A capacity provider strategy consists of one or more capacity providers along with the base and weight to assign to them. A capacity provider must be associated with the cluster to be used in a capacity provider strategy. The PutClusterCapacityProviders API is used to associate a capacity provider with a cluster. Only capacity providers with an ACTIVE or UPDATING status can be used.

If a capacityProviderStrategy is specified, the launchType parameter must be omitted. If no capacityProviderStrategy or launchType is specified, the defaultCapacityProviderStrategy for the cluster is used.

If specifying a capacity provider that uses an Auto Scaling group, the capacity provider must already be created. New capacity providers can be created with the CreateCapacityProvider API operation.

To use a AWS Fargate capacity provider, specify either the FARGATE or FARGATE_SPOT capacity providers. The AWS Fargate capacity providers are available to all accounts and only need to be associated with a cluster to be used.

The PutClusterCapacityProviders API operation is used to update the list of available capacity providers for a cluster after the cluster is created.

pub fn set_capacity_provider_strategy(
self,
input: Option<Vec<CapacityProviderStrategyItem>>
) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

The platform version that the tasks in the task set should use. A platform version is specified only for tasks using the Fargate launch type. If one isn't specified, the LATEST platform version is used by default.

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn scale(self, input: Scale) -> Self

A floating-point percentage of the desired number of tasks to place and keep running in the task set.

pub fn set_scale(self, input: Option<Scale>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. When a service is deleted, the tags are deleted as well.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for AWS use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self