Struct aws_sdk_ecs::client::fluent_builders::ListAccountSettings

pub struct ListAccountSettings { /* fields omitted */ }

Implementations

impl ListAccountSettings

pub async fn send(
self
) -> Result<ListAccountSettingsOutput, SdkError<ListAccountSettingsError>>

pub fn name(self, inp: SettingName) -> Self

The name of the account setting you want to list the settings for.

pub fn set_name(self, inp: Option<SettingName>) -> Self

pub fn value(self, inp: impl Into<String>) -> Self

The value of the account settings with which to filter results. You must also specify an account setting name to use this parameter.

pub fn set_value(self, inp: Option<String>) -> Self

pub fn principal_arn(self, inp: impl Into<String>) -> Self

The ARN of the principal, which can be an IAM user, IAM role, or the root user. If this field is omitted, the account settings are listed only for the authenticated user.

Federated users assume the account setting of the root user and can't have explicit account settings set for them.

pub fn set_principal_arn(self, inp: Option<String>) -> Self

pub fn effective_settings(self, inp: bool) -> Self

Specifies whether to return the effective settings. If true, the account settings for the root user or the default setting for the principalArn are returned. If false, the account settings for the principalArn are returned if they are set. Otherwise, no account settings are returned.

pub fn set_effective_settings(self, inp: bool) -> Self

pub fn next_token(self, inp: impl Into<String>) -> Self

The nextToken value returned from a ListAccountSettings request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

pub fn set_next_token(self, inp: Option<String>) -> Self

pub fn max_results(self, inp: i32) -> Self

The maximum number of account setting results returned by ListAccountSettings in paginated output. When this parameter is used, ListAccountSettings only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListAccountSettings request with the returned nextToken value. This value can be between 1 and 10. If this parameter is not used, then ListAccountSettings returns up to 10 results and a nextToken value if applicable.

pub fn set_max_results(self, inp: i32) -> Self