Struct aws_sdk_ecs::client::fluent_builders::ListTasks

pub struct ListTasks<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> ListTasks<C>

pub async fn send(self) -> Result<ListTasksOutput, SdkError<ListTasksError>> where
C: SmithyConnector,

pub fn cluster(self, input: impl Into<String>) -> Self

The short name or full Amazon Resource Name (ARN) of the cluster to use when filtering the ListTasks results. If you do not specify a cluster, the default cluster is assumed.

pub fn set_cluster(self, input: Option<String>) -> Self

pub fn container_instance(self, input: impl Into<String>) -> Self

The container instance ID or full ARN of the container instance to use when filtering the ListTasks results. Specifying a containerInstance limits the results to tasks that belong to that container instance.

pub fn set_container_instance(self, input: Option<String>) -> Self

pub fn family(self, input: impl Into<String>) -> Self

The name of the task definition family to use when filtering the ListTasks results. Specifying a family limits the results to tasks that belong to that family.

pub fn set_family(self, input: Option<String>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

The nextToken value returned from a ListTasks request indicating that more results are available to fulfill the request and further calls will be needed. If maxResults was provided, it is possible the number of results to be fewer than maxResults.

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn max_results(self, input: i32) -> Self

The maximum number of task results returned by ListTasks in paginated output. When this parameter is used, ListTasks only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTasks request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTasks returns up to 100 results and a nextToken value if applicable.

pub fn set_max_results(self, input: Option<i32>) -> Self

pub fn started_by(self, input: impl Into<String>) -> Self

The startedBy value with which to filter the task results. Specifying a startedBy value limits the results to tasks that were started with that value.

pub fn set_started_by(self, input: Option<String>) -> Self

pub fn service_name(self, input: impl Into<String>) -> Self

The name of the service to use when filtering the ListTasks results. Specifying a serviceName limits the results to tasks that belong to that service.

pub fn set_service_name(self, input: Option<String>) -> Self

pub fn desired_status(self, input: DesiredStatus) -> Self

The task desired status to use when filtering the ListTasks results. Specifying a desiredStatus of STOPPED limits the results to tasks that Amazon ECS has set the desired status to STOPPED. This can be useful for debugging tasks that are not starting properly or have died or finished. The default status filter is RUNNING, which shows tasks that Amazon ECS has set the desired status to RUNNING.

Although you can filter results based on a desired status of PENDING, this does not return any results. Amazon ECS never sets the desired status of a task to that value (only a task's lastStatus may have a value of PENDING).

pub fn set_desired_status(self, input: Option<DesiredStatus>) -> Self

pub fn launch_type(self, input: LaunchType) -> Self

The launch type to use when filtering the ListTasks results.

pub fn set_launch_type(self, input: Option<LaunchType>) -> Self