Struct aws_sdk_ecs::client::fluent_builders::SubmitTaskStateChange

pub struct SubmitTaskStateChange { /* fields omitted */ }

Implementations

impl SubmitTaskStateChange

pub async fn send(
self
) -> Result<SubmitTaskStateChangeOutput, SdkError<SubmitTaskStateChangeError>>

pub fn cluster(self, inp: impl Into<String>) -> Self

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task.

pub fn set_cluster(self, inp: Option<String>) -> Self

pub fn task(self, inp: impl Into<String>) -> Self

The task ID or full ARN of the task in the state change request.

pub fn set_task(self, inp: Option<String>) -> Self

pub fn status(self, inp: impl Into<String>) -> Self

The status of the state change request.

pub fn set_status(self, inp: Option<String>) -> Self

pub fn reason(self, inp: impl Into<String>) -> Self

The reason for the state change request.

pub fn set_reason(self, inp: Option<String>) -> Self

pub fn containers(self, inp: impl Into<ContainerStateChange>) -> Self

Any containers associated with the state change request.

pub fn set_containers(self, inp: Option<Vec<ContainerStateChange>>) -> Self

pub fn attachments(self, inp: impl Into<AttachmentStateChange>) -> Self

Any attachments associated with the state change request.

pub fn set_attachments(self, inp: Option<Vec<AttachmentStateChange>>) -> Self

pub fn managed_agents(self, inp: impl Into<ManagedAgentStateChange>) -> Self

The details for the managed agent associated with the task.

pub fn set_managed_agents(
self,
inp: Option<Vec<ManagedAgentStateChange>>
) -> Self

pub fn pull_started_at(self, inp: Instant) -> Self

The Unix timestamp for when the container image pull began.

pub fn set_pull_started_at(self, inp: Option<Instant>) -> Self

pub fn pull_stopped_at(self, inp: Instant) -> Self

The Unix timestamp for when the container image pull completed.

pub fn set_pull_stopped_at(self, inp: Option<Instant>) -> Self

pub fn execution_stopped_at(self, inp: Instant) -> Self

The Unix timestamp for when the task execution stopped.

pub fn set_execution_stopped_at(self, inp: Option<Instant>) -> Self