Struct aws_sdk_ecs::client::fluent_builders::UpdateService

pub struct UpdateService<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> UpdateService<C>

pub async fn send(
self
) -> Result<UpdateServiceOutput, SdkError<UpdateServiceError>> where
C: SmithyConnector,

pub fn cluster(self, input: impl Into<String>) -> Self

The short name or full Amazon Resource Name (ARN) of the cluster that your service is running on. If you do not specify a cluster, the default cluster is assumed.

pub fn set_cluster(self, input: Option<String>) -> Self

pub fn service(self, input: impl Into<String>) -> Self

The name of the service to update.

pub fn set_service(self, input: Option<String>) -> Self

pub fn desired_count(self, input: i32) -> Self

The number of instantiations of the task to place and keep running in your service.

pub fn set_desired_count(self, input: Option<i32>) -> Self

pub fn task_definition(self, input: impl Into<String>) -> Self

The family and revision (family:revision) or full ARN of the task definition to run in your service. If a revision is not specified, the latest ACTIVE revision is used. If you modify the task definition with UpdateService, Amazon ECS spawns a task with the new version of the task definition and then stops an old task after the new version is running.

pub fn set_task_definition(self, input: Option<String>) -> Self

pub fn capacity_provider_strategy(
self,
inp: impl Into<CapacityProviderStrategyItem>
) -> Self

The capacity provider strategy to update the service to use.

If the service is using the default capacity provider strategy for the cluster, the service can be updated to use one or more capacity providers as opposed to the default capacity provider strategy. However, when a service is using a capacity provider strategy that is not the default capacity provider strategy, the service cannot be updated to use the cluster's default capacity provider strategy.

A capacity provider strategy consists of one or more capacity providers along with the base and weight to assign to them. A capacity provider must be associated with the cluster to be used in a capacity provider strategy. The PutClusterCapacityProviders API is used to associate a capacity provider with a cluster. Only capacity providers with an ACTIVE or UPDATING status can be used.

If specifying a capacity provider that uses an Auto Scaling group, the capacity provider must already be created. New capacity providers can be created with the CreateCapacityProvider API operation.

To use a AWS Fargate capacity provider, specify either the FARGATE or FARGATE_SPOT capacity providers. The AWS Fargate capacity providers are available to all accounts and only need to be associated with a cluster to be used.

The PutClusterCapacityProviders API operation is used to update the list of available capacity providers for a cluster after the cluster is created.

pub fn set_capacity_provider_strategy(
self,
input: Option<Vec<CapacityProviderStrategyItem>>
) -> Self

pub fn deployment_configuration(self, input: DeploymentConfiguration) -> Self

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

pub fn set_deployment_configuration(
self,
input: Option<DeploymentConfiguration>
) -> Self

pub fn network_configuration(self, input: NetworkConfiguration) -> Self

An object representing the network configuration for the service.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

pub fn placement_constraints(self, inp: impl Into<PlacementConstraint>) -> Self

An array of task placement constraint objects to update the service to use. If no value is specified, the existing placement constraints for the service will remain unchanged. If this value is specified, it will override any existing placement constraints defined for the service. To remove all existing placement constraints, specify an empty array.

You can specify a maximum of 10 constraints per task (this limit includes constraints in the task definition and those specified at runtime).

pub fn set_placement_constraints(
self,
input: Option<Vec<PlacementConstraint>>
) -> Self

pub fn placement_strategy(self, inp: impl Into<PlacementStrategy>) -> Self

The task placement strategy objects to update the service to use. If no value is specified, the existing placement strategy for the service will remain unchanged. If this value is specified, it will override the existing placement strategy defined for the service. To remove an existing placement strategy, specify an empty object.

You can specify a maximum of five strategy rules per service.

pub fn set_placement_strategy(
self,
input: Option<Vec<PlacementStrategy>>
) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

The platform version on which your tasks in the service are running. A platform version is only specified for tasks using the Fargate launch type. If a platform version is not specified, the LATEST platform version is used by default. For more information, see AWS Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn force_new_deployment(self, input: bool) -> Self

Whether to force a new deployment of the service. Deployments are not forced by default. You can use this option to trigger a new deployment with no service definition changes. For example, you can update a service's tasks to use a newer Docker image with the same image/tag combination (my_image:latest) or to roll Fargate tasks onto a newer platform version.

pub fn set_force_new_deployment(self, input: Option<bool>) -> Self

pub fn health_check_grace_period_seconds(self, input: i32) -> Self

The period of time, in seconds, that the Amazon ECS service scheduler should ignore unhealthy Elastic Load Balancing target health checks after a task has first started. This is only valid if your service is configured to use a load balancer. If your service's tasks take a while to start and respond to Elastic Load Balancing health checks, you can specify a health check grace period of up to 2,147,483,647 seconds. During that time, the Amazon ECS service scheduler ignores the Elastic Load Balancing health check status. This grace period can prevent the ECS service scheduler from marking tasks as unhealthy and stopping them before they have time to come up.

pub fn set_health_check_grace_period_seconds(self, input: Option<i32>) -> Self

pub fn enable_execute_command(self, input: bool) -> Self

If true, this enables execute command functionality on all task containers.

If you do not want to override the value that was set when the service was created, you can set this to null when performing this action.

pub fn set_enable_execute_command(self, input: Option<bool>) -> Self