Struct aws_sdk_ecs::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonEC2ContainerServiceV20141113.

This client allows ergonomic access to a AmazonEC2ContainerServiceV20141113-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_capacity_provider(&self) -> CreateCapacityProvider<C, M, R>

pub fn create_cluster(&self) -> CreateCluster<C, M, R>

pub fn create_service(&self) -> CreateService<C, M, R>

pub fn create_task_set(&self) -> CreateTaskSet<C, M, R>

pub fn delete_account_setting(&self) -> DeleteAccountSetting<C, M, R>

pub fn delete_attributes(&self) -> DeleteAttributes<C, M, R>

pub fn delete_capacity_provider(&self) -> DeleteCapacityProvider<C, M, R>

pub fn delete_cluster(&self) -> DeleteCluster<C, M, R>

pub fn delete_service(&self) -> DeleteService<C, M, R>

pub fn delete_task_set(&self) -> DeleteTaskSet<C, M, R>

pub fn deregister_container_instance(
&self
) -> DeregisterContainerInstance<C, M, R>

pub fn deregister_task_definition(&self) -> DeregisterTaskDefinition<C, M, R>

pub fn describe_capacity_providers(&self) -> DescribeCapacityProviders<C, M, R>

pub fn describe_clusters(&self) -> DescribeClusters<C, M, R>

pub fn describe_container_instances(
&self
) -> DescribeContainerInstances<C, M, R>

pub fn describe_services(&self) -> DescribeServices<C, M, R>

pub fn describe_task_definition(&self) -> DescribeTaskDefinition<C, M, R>

pub fn describe_tasks(&self) -> DescribeTasks<C, M, R>

pub fn describe_task_sets(&self) -> DescribeTaskSets<C, M, R>

pub fn discover_poll_endpoint(&self) -> DiscoverPollEndpoint<C, M, R>

pub fn execute_command(&self) -> ExecuteCommand<C, M, R>

pub fn list_account_settings(&self) -> ListAccountSettings<C, M, R>

pub fn list_attributes(&self) -> ListAttributes<C, M, R>

pub fn list_clusters(&self) -> ListClusters<C, M, R>

pub fn list_container_instances(&self) -> ListContainerInstances<C, M, R>

pub fn list_services(&self) -> ListServices<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_task_definition_families(
&self
) -> ListTaskDefinitionFamilies<C, M, R>

pub fn list_task_definitions(&self) -> ListTaskDefinitions<C, M, R>

pub fn list_tasks(&self) -> ListTasks<C, M, R>

pub fn put_account_setting(&self) -> PutAccountSetting<C, M, R>

pub fn put_account_setting_default(&self) -> PutAccountSettingDefault<C, M, R>

pub fn put_attributes(&self) -> PutAttributes<C, M, R>

pub fn put_cluster_capacity_providers(
&self
) -> PutClusterCapacityProviders<C, M, R>

pub fn register_container_instance(&self) -> RegisterContainerInstance<C, M, R>

pub fn register_task_definition(&self) -> RegisterTaskDefinition<C, M, R>

pub fn run_task(&self) -> RunTask<C, M, R>

pub fn start_task(&self) -> StartTask<C, M, R>

pub fn stop_task(&self) -> StopTask<C, M, R>

pub fn submit_attachment_state_changes(
&self
) -> SubmitAttachmentStateChanges<C, M, R>

pub fn submit_container_state_change(
&self
) -> SubmitContainerStateChange<C, M, R>

pub fn submit_task_state_change(&self) -> SubmitTaskStateChange<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_capacity_provider(&self) -> UpdateCapacityProvider<C, M, R>

pub fn update_cluster(&self) -> UpdateCluster<C, M, R>

pub fn update_cluster_settings(&self) -> UpdateClusterSettings<C, M, R>

pub fn update_container_agent(&self) -> UpdateContainerAgent<C, M, R>

pub fn update_container_instances_state(
&self
) -> UpdateContainerInstancesState<C, M, R>

pub fn update_service(&self) -> UpdateService<C, M, R>

pub fn update_service_primary_task_set(
&self
) -> UpdateServicePrimaryTaskSet<C, M, R>

pub fn update_task_set(&self) -> UpdateTaskSet<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self