Module aws_sdk_ecs::error

Modules

access_denied_exception

See AccessDeniedException

attribute_limit_exceeded_exception

See AttributeLimitExceededException

blocked_exception

See BlockedException

client_exception

See ClientException

cluster_contains_container_instances_exception

See ClusterContainsContainerInstancesException

cluster_contains_services_exception

See ClusterContainsServicesException

cluster_contains_tasks_exception

See ClusterContainsTasksException

cluster_not_found_exception

See ClusterNotFoundException

invalid_parameter_exception

See InvalidParameterException

limit_exceeded_exception

See LimitExceededException

missing_version_exception

See MissingVersionException

no_update_available_exception

See NoUpdateAvailableException

platform_task_definition_incompatibility_exception

See PlatformTaskDefinitionIncompatibilityException

platform_unknown_exception

See PlatformUnknownException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

server_exception

See ServerException

service_not_active_exception

See ServiceNotActiveException

service_not_found_exception

See ServiceNotFoundException

target_not_connected_exception

See TargetNotConnectedException

target_not_found_exception

See TargetNotFoundException

task_set_not_found_exception

See TaskSetNotFoundException

unsupported_feature_exception

See UnsupportedFeatureException

update_in_progress_exception

See UpdateInProgressException

Structs

AccessDeniedException

You do not have authorization to perform the requested action.

AttributeLimitExceededException

You can apply up to 10 custom attributes per resource. You can view the attributes of

BlockedException

Your AWS account has been blocked. For more information, contact AWS Support.

ClientException

These errors are usually caused by a client action, such as using an action or

ClusterContainsContainerInstancesException

You cannot delete a cluster that has registered container instances. First, deregister

ClusterContainsServicesException

You cannot delete a cluster that contains services. First, update the service to

ClusterContainsTasksException

You cannot delete a cluster that has active tasks.

ClusterNotFoundException

The specified cluster could not be found. You can view your available clusters with

CreateCapacityProviderError

CreateClusterError

CreateServiceError

CreateTaskSetError

DeleteAccountSettingError

DeleteAttributesError

DeleteCapacityProviderError

DeleteClusterError

DeleteServiceError

DeleteTaskSetError

DeregisterContainerInstanceError

DeregisterTaskDefinitionError

DescribeCapacityProvidersError

DescribeClustersError

DescribeContainerInstancesError

DescribeServicesError

DescribeTaskDefinitionError

DescribeTaskSetsError

DescribeTasksError

DiscoverPollEndpointError

ExecuteCommandError

InvalidParameterException

The specified parameter is invalid. Review the available parameters for the API

LimitExceededException

The limit for the resource has been exceeded.

ListAccountSettingsError

ListAttributesError

ListClustersError

ListContainerInstancesError

ListServicesError

ListTagsForResourceError

ListTaskDefinitionFamiliesError

ListTaskDefinitionsError

ListTasksError

MissingVersionException

Amazon ECS is unable to determine the current version of the Amazon ECS container agent on the

NoUpdateAvailableException

There is no update available for this Amazon ECS container agent. This could be because the

PlatformTaskDefinitionIncompatibilityException

The specified platform version does not satisfy the task definition's required

PlatformUnknownException

The specified platform version does not exist.

PutAccountSettingDefaultError

PutAccountSettingError

PutAttributesError

PutClusterCapacityProvidersError

RegisterContainerInstanceError

RegisterTaskDefinitionError

ResourceInUseException

The specified resource is in-use and cannot be removed.

ResourceNotFoundException

The specified resource could not be found.

RunTaskError

ServerException

These errors are usually caused by a server issue.

ServiceNotActiveException

The specified service is not active. You can't update a service that is inactive. If

ServiceNotFoundException

The specified service could not be found. You can view your available services with

StartTaskError

StopTaskError

SubmitAttachmentStateChangesError

SubmitContainerStateChangeError

SubmitTaskStateChangeError

TagResourceError

TargetNotConnectedException

The target container is not properly configured with the execute command agent or the

TargetNotFoundException

The specified target could not be found. You can view your available container

TaskSetNotFoundException

The specified task set could not be found. You can view your available task sets with

UnsupportedFeatureException

The specified task is not supported in this Region.

UntagResourceError

UpdateCapacityProviderError

UpdateClusterError

UpdateClusterSettingsError

UpdateContainerAgentError

UpdateContainerInstancesStateError

UpdateInProgressException

There is already a current Amazon ECS container agent update in progress on the specified

UpdateServiceError

UpdateServicePrimaryTaskSetError

UpdateTaskSetError

Enums

CreateCapacityProviderErrorKind

CreateClusterErrorKind

CreateServiceErrorKind

CreateTaskSetErrorKind

DeleteAccountSettingErrorKind

DeleteAttributesErrorKind

DeleteCapacityProviderErrorKind

DeleteClusterErrorKind

DeleteServiceErrorKind

DeleteTaskSetErrorKind

DeregisterContainerInstanceErrorKind

DeregisterTaskDefinitionErrorKind

DescribeCapacityProvidersErrorKind

DescribeClustersErrorKind

DescribeContainerInstancesErrorKind

DescribeServicesErrorKind

DescribeTaskDefinitionErrorKind

DescribeTaskSetsErrorKind

DescribeTasksErrorKind

DiscoverPollEndpointErrorKind

ExecuteCommandErrorKind

ListAccountSettingsErrorKind

ListAttributesErrorKind

ListClustersErrorKind

ListContainerInstancesErrorKind

ListServicesErrorKind

ListTagsForResourceErrorKind

ListTaskDefinitionFamiliesErrorKind

ListTaskDefinitionsErrorKind

ListTasksErrorKind

PutAccountSettingDefaultErrorKind

PutAccountSettingErrorKind

PutAttributesErrorKind

PutClusterCapacityProvidersErrorKind

RegisterContainerInstanceErrorKind

RegisterTaskDefinitionErrorKind

RunTaskErrorKind

StartTaskErrorKind

StopTaskErrorKind

SubmitAttachmentStateChangesErrorKind

SubmitContainerStateChangeErrorKind

SubmitTaskStateChangeErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateCapacityProviderErrorKind

UpdateClusterErrorKind

UpdateClusterSettingsErrorKind

UpdateContainerAgentErrorKind

UpdateContainerInstancesStateErrorKind

UpdateServiceErrorKind

UpdateServicePrimaryTaskSetErrorKind

UpdateTaskSetErrorKind