Struct aws_sdk_ecs::input::run_task_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for RunTaskInput

Implementations

impl Builder

pub fn capacity_provider_strategy(
self,
input: impl Into<CapacityProviderStrategyItem>
) -> Self

pub fn set_capacity_provider_strategy(
self,
input: Option<Vec<CapacityProviderStrategyItem>>
) -> Self

pub fn cluster(self, input: impl Into<String>) -> Self

The short name or full Amazon Resource Name (ARN) of the cluster on which to run your task. If you do not specify a cluster, the default cluster is assumed.

pub fn set_cluster(self, input: Option<String>) -> Self

pub fn count(self, input: i32) -> Self

The number of instantiations of the specified task to place on your cluster. You can specify up to 10 tasks per call.

pub fn set_count(self, input: Option<i32>) -> Self

pub fn enable_ecs_managed_tags(self, input: bool) -> Self

Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

pub fn set_enable_ecs_managed_tags(self, input: Option<bool>) -> Self

pub fn enable_execute_command(self, input: bool) -> Self

Whether or not to enable the execute command functionality for the containers in this task. If true, this enables execute command functionality on all containers in the task.

pub fn set_enable_execute_command(self, input: Option<bool>) -> Self

pub fn group(self, input: impl Into<String>) -> Self

The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

pub fn set_group(self, input: Option<String>) -> Self

pub fn launch_type(self, input: LaunchType) -> Self

The infrastructure on which to run your standalone task. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

The FARGATE launch type runs your tasks on Fargate On-Demand infrastructure.

Fargate Spot infrastructure is available for use but a capacity provider strategy must be used. For more information, see Fargate capacity providers in the Amazon ECS User Guide for Fargate.

The EC2 launch type runs your tasks on Amazon EC2 instances registered to your cluster.

The EXTERNAL launch type runs your tasks on your on-premise server or virtual machine (VM) capacity registered to your cluster.

A task can use either a launch type or a capacity provider strategy. If a launchType is specified, the capacityProviderStrategy parameter must be omitted.

When you use cluster auto scaling, you must specify capacityProviderStrategy and not launchType.

pub fn set_launch_type(self, input: Option<LaunchType>) -> Self

pub fn network_configuration(self, input: NetworkConfiguration) -> Self

The network configuration for the task. This parameter is required for task definitions that use the awsvpc network mode to receive their own elastic network interface, and it is not supported for other network modes. For more information, see Task networking in the Amazon Elastic Container Service Developer Guide.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

pub fn overrides(self, input: TaskOverride) -> Self

A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

A total of 8192 characters are allowed for overrides. This limit includes the JSON formatting characters of the override structure.

pub fn set_overrides(self, input: Option<TaskOverride>) -> Self

pub fn placement_constraints(
self,
input: impl Into<PlacementConstraint>
) -> Self

pub fn set_placement_constraints(
self,
input: Option<Vec<PlacementConstraint>>
) -> Self

pub fn placement_strategy(self, input: impl Into<PlacementStrategy>) -> Self

pub fn set_placement_strategy(
self,
input: Option<Vec<PlacementStrategy>>
) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

The platform version the task should use. A platform version is only specified for tasks hosted on Fargate. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate platform versions in the Amazon Elastic Container Service Developer Guide.

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn propagate_tags(self, input: PropagateTags) -> Self

Specifies whether to propagate the tags from the task definition to the task. If no value is specified, the tags are not propagated. Tags can only be propagated to the task during task creation. To add tags to a task after task creation, use the TagResource API action.

An error will be received if you specify the SERVICE option when running a task.

pub fn set_propagate_tags(self, input: Option<PropagateTags>) -> Self

pub fn reference_id(self, input: impl Into<String>) -> Self

The reference ID to use for the task. The reference ID can have a maximum length of 1024 characters.

pub fn set_reference_id(self, input: Option<String>) -> Self

pub fn started_by(self, input: impl Into<String>) -> Self

An optional tag specified when a task is started. For example, if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

pub fn set_started_by(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn task_definition(self, input: impl Into<String>) -> Self

The family and revision (family:revision) or full ARN of the task definition to run. If a revision is not specified, the latest ACTIVE revision is used.

pub fn set_task_definition(self, input: Option<String>) -> Self

pub fn build(self) -> Result<RunTaskInput, BuildError>

Consumes the builder and constructs a RunTaskInput