Module aws_sdk_ecs::model

Modules

attachment

See Attachment

attachment_state_change

See AttachmentStateChange

attribute

See Attribute

auto_scaling_group_provider

See AutoScalingGroupProvider

auto_scaling_group_provider_update

See AutoScalingGroupProviderUpdate

aws_vpc_configuration

See AwsVpcConfiguration

capacity_provider

See CapacityProvider

capacity_provider_strategy_item

See CapacityProviderStrategyItem

cluster

See Cluster

cluster_configuration

See ClusterConfiguration

cluster_setting

See ClusterSetting

container

See Container

container_definition

See ContainerDefinition

container_dependency

See ContainerDependency

container_instance

See ContainerInstance

container_override

See ContainerOverride

container_state_change

See ContainerStateChange

deployment

See Deployment

deployment_circuit_breaker

See DeploymentCircuitBreaker

deployment_configuration

See DeploymentConfiguration

deployment_controller

See DeploymentController

device

See Device

docker_volume_configuration

See DockerVolumeConfiguration

efs_authorization_config

See EfsAuthorizationConfig

efs_volume_configuration

See EfsVolumeConfiguration

environment_file

See EnvironmentFile

ephemeral_storage

See EphemeralStorage

execute_command_configuration

See ExecuteCommandConfiguration

execute_command_log_configuration

See ExecuteCommandLogConfiguration

f_sx_windows_file_server_authorization_config

See FSxWindowsFileServerAuthorizationConfig

f_sx_windows_file_server_volume_configuration

See FSxWindowsFileServerVolumeConfiguration

failure

See Failure

firelens_configuration

See FirelensConfiguration

health_check

See HealthCheck

host_entry

See HostEntry

host_volume_properties

See HostVolumeProperties

inference_accelerator

See InferenceAccelerator

inference_accelerator_override

See InferenceAcceleratorOverride

kernel_capabilities

See KernelCapabilities

key_value_pair

See KeyValuePair

linux_parameters

See LinuxParameters

load_balancer

See LoadBalancer

log_configuration

See LogConfiguration

managed_agent

See ManagedAgent

managed_agent_state_change

See ManagedAgentStateChange

managed_scaling

See ManagedScaling

mount_point

See MountPoint

network_binding

See NetworkBinding

network_configuration

See NetworkConfiguration

network_interface

See NetworkInterface

placement_constraint

See PlacementConstraint

placement_strategy

See PlacementStrategy

platform_device

See PlatformDevice

port_mapping

See PortMapping

proxy_configuration

See ProxyConfiguration

repository_credentials

See RepositoryCredentials

resource

See Resource

resource_requirement

See ResourceRequirement

scale

See Scale

secret

See Secret

service

See Service

service_event

See ServiceEvent

service_registry

See ServiceRegistry

session

See Session

setting

See Setting

system_control

See SystemControl

tag

See Tag

task

See Task

task_definition

See TaskDefinition

task_definition_placement_constraint

See TaskDefinitionPlacementConstraint

task_override

See TaskOverride

task_set

See TaskSet

tmpfs

See Tmpfs

ulimit

See Ulimit

version_info

See VersionInfo

volume

See Volume

volume_from

See VolumeFrom

Structs

Attachment

An object representing a container instance or task attachment.

AttachmentStateChange

An object representing a change in state for a task attachment.

Attribute

An attribute is a name-value pair associated with an Amazon ECS object. Attributes enable

AutoScalingGroupProvider

The details of the Auto Scaling group for the capacity provider.

AutoScalingGroupProviderUpdate

The details of the Auto Scaling group capacity provider to update.

AwsVpcConfiguration

An object representing the networking details for a task or service.

CapacityProvider

The details of a capacity provider.

CapacityProviderStrategyItem

The details of a capacity provider strategy. A capacity provider strategy can be set

Cluster

A regional grouping of one or more container instances on which you can run task

ClusterConfiguration

The execute command configuration for the cluster.

ClusterSetting

The settings to use when creating a cluster. This parameter is used to enable CloudWatch

Container

A Docker container that is part of a task.

ContainerDefinition

Container definitions are used in task definitions to describe the different

ContainerDependency

The dependencies defined for container startup and shutdown. A container can contain

ContainerInstance

An EC2 instance that is running the Amazon ECS agent and has been registered with a

ContainerOverride

The overrides that should be sent to a container. An empty container override can be

ContainerStateChange

An object representing a change in state for a container.

Deployment

The details of an Amazon ECS service deployment. This is used only when a service uses the

DeploymentCircuitBreaker

DeploymentConfiguration

Optional deployment parameters that control how many tasks run during a deployment and

DeploymentController

The deployment controller to use for the service. For more information, see Amazon ECS Deployment Types in the Amazon Elastic Container Service Developer Guide.

Device

An object representing a container instance host device.

DockerVolumeConfiguration

This parameter is specified when you are using Docker volumes. Docker volumes are only

EfsAuthorizationConfig

The authorization configuration details for the Amazon EFS file system.

EfsVolumeConfiguration

This parameter is specified when you are using an Amazon Elastic File System file system for task

EnvironmentFile

A list of files containing the environment variables to pass to a container. You can

EphemeralStorage

The amount of ephemeral storage to allocate for the task. This parameter is used to

ExecuteCommandConfiguration

The details of the execute command configuration.

ExecuteCommandLogConfiguration

The log configuration for the results of the execute command actions. The logs can be

FSxWindowsFileServerAuthorizationConfig

The authorization configuration details for Amazon FSx for Windows File Server file system. See FSxWindowsFileServerVolumeConfiguration in the Amazon Elastic

FSxWindowsFileServerVolumeConfiguration

This parameter is specified when you are using Amazon FSx for Windows File Server file system for task

Failure

A failed resource. For a list of common causes, see API failure

FirelensConfiguration

The FireLens configuration for the container. This is used to specify and configure a

HealthCheck

An object representing a container health check. Health check parameters that are

HostEntry

Hostnames and IP address entries that are added to the /etc/hosts file of

HostVolumeProperties

Details on a container instance bind mount host volume.

InferenceAccelerator

Details on a Elastic Inference accelerator. For more information, see Working with

InferenceAcceleratorOverride

Details on an Elastic Inference accelerator task override. This parameter is used to

KernelCapabilities

The Linux capabilities for the container that are added to or dropped from the default

KeyValuePair

A key-value pair object.

LinuxParameters

Linux-specific options that are applied to the container, such as Linux KernelCapabilities.

LoadBalancer

The load balancer configuration to use with a service or task set.

LogConfiguration

The log configuration for the container. This parameter maps to LogConfig

ManagedAgent

Details about the managed agent status for the container.

ManagedAgentStateChange

An object representing a change in state for a managed agent.

ManagedScaling

The managed scaling settings for the Auto Scaling group capacity provider.

MountPoint

Details on a volume mount point that is used in a container definition.

NetworkBinding

Details on the network bindings between a container and its host container instance.

NetworkConfiguration

An object representing the network configuration for a task or service.

NetworkInterface

An object representing the elastic network interface for tasks that use the

PlacementConstraint

An object representing a constraint on task placement. For more information, see

PlacementStrategy

The task placement strategy for a task or service. For more information, see Task Placement Strategies in the

PlatformDevice

The devices that are available on the container instance. The only supported device

PortMapping

Port mappings allow containers to access ports on the host container instance to send

ProxyConfiguration

The configuration details for the App Mesh proxy.

RepositoryCredentials

The repository credentials for private registry authentication.

Resource

Describes the resources available for a container instance.

ResourceRequirement

The type and amount of a resource to assign to a container. The supported resource

Scale

A floating-point percentage of the desired number of tasks to place and keep running

Secret

An object representing the secret to expose to your container. Secrets can be exposed

Service

Details on a service within a cluster

ServiceEvent

Details on an event associated with a service.

ServiceRegistry

Details of the service registry.

Session

The details of the execute command session.

Setting

The current account setting for a resource.

SystemControl

A list of namespaced kernel parameters to set in the container. This parameter maps to

Tag

The metadata that you apply to a resource to help you categorize and organize them.

Task

Details on a task in a cluster.

TaskDefinition

The details of a task definition which describes the container and volume definitions

TaskDefinitionPlacementConstraint

An object representing a constraint on task placement in the task definition. For more

TaskOverride

The overrides associated with a task.

TaskSet

Information about a set of Amazon ECS tasks in either an AWS CodeDeploy or an EXTERNAL

Tmpfs

The container path, mount options, and size of the tmpfs mount.

Ulimit

The ulimit settings to pass to the container.

VersionInfo

The Docker and Amazon ECS container agent version information about a container

Volume

A data volume used in a task definition. For tasks that use the Amazon Elastic File

VolumeFrom

Details on a data volume from another container in the same task definition.

Enums

AgentUpdateStatus

AssignPublicIp

CapacityProviderField

CapacityProviderStatus

CapacityProviderUpdateStatus

ClusterField

ClusterSettingName

Compatibility

Connectivity

ContainerCondition

ContainerInstanceField

ContainerInstanceStatus

DeploymentControllerType

DeploymentRolloutState

DesiredStatus

DeviceCgroupPermission

EfsAuthorizationConfigIam

EfsTransitEncryption

EnvironmentFileType

ExecuteCommandLogging

FirelensConfigurationType

HealthStatus

NOTE: HealthStatus::Unknown has been renamed to ::UnknownValue.

IpcMode

LaunchType

LogDriver

ManagedAgentName

ManagedScalingStatus

ManagedTerminationProtection

NetworkMode

PidMode

PlacementConstraintType

PlacementStrategyType

PlatformDeviceType

PropagateTags

ProxyConfigurationType

ResourceType

ScaleUnit

SchedulingStrategy

Scope

ServiceField

SettingName

SortOrder

StabilityStatus

TargetType

TaskDefinitionFamilyStatus

TaskDefinitionField

TaskDefinitionPlacementConstraintType

TaskDefinitionStatus

TaskField

TaskSetField

TaskStopCode

TransportProtocol

UlimitName