Struct aws_sdk_ecs::model::container_definition::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ContainerDefinition

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of a container. If you are linking multiple containers together in a task definition, the name of one container can be entered in the links of another container to connect the containers. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. This parameter maps to name in the Create a container section of the Docker Remote API and the --name option to docker run.

pub fn set_name(self, input: Option<String>) -> Self

pub fn image(self, input: impl Into<String>) -> Self

The image used to start a container. This string is passed directly to the Docker daemon. Images in the Docker Hub registry are available by default. Other repositories are specified with either repository-url/image:tag or repository-url/image@digest . Up to 255 letters (uppercase and lowercase), numbers, hyphens, underscores, colons, periods, forward slashes, and number signs are allowed. This parameter maps to Image in the Create a container section of the Docker Remote API and the IMAGE parameter of docker run.

When a new task starts, the Amazon ECS container agent pulls the latest version of the specified image and tag for the container to use. However, subsequent updates to a repository image are not propagated to already running tasks.

Images in Amazon ECR repositories can be specified by either using the full registry/repository:tag or registry/repository@digest. For example, 012345678910.dkr.ecr..amazonaws.com/:latest or 012345678910.dkr.ecr..amazonaws.com/@sha256:94afd1f2e64d908bc90dbca0035a5b567EXAMPLE.

Images in official repositories on Docker Hub use a single name (for example, ubuntu or mongo).

Images in other repositories on Docker Hub are qualified with an organization name (for example, amazon/amazon-ecs-agent).

Images in other online repositories are qualified further by a domain name (for example, quay.io/assemblyline/ubuntu).

pub fn set_image(self, input: Option<String>) -> Self

pub fn repository_credentials(self, input: RepositoryCredentials) -> Self

The private repository authentication credentials to use.

pub fn set_repository_credentials(
self,
input: Option<RepositoryCredentials>
) -> Self

pub fn cpu(self, input: i32) -> Self

The number of cpu units reserved for the container. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run.

This field is optional for tasks using the Fargate launch type, and the only requirement is that the total amount of CPU reserved for all containers within a task be lower than the task-level cpu value.

You can determine the number of CPU units that are available per EC2 instance type by multiplying the vCPUs listed for that instance type on the Amazon EC2 Instances detail page by 1,024.

Linux containers share unallocated CPU units with other containers on the container instance with the same ratio as their allocated amount. For example, if you run a single-container task on a single-core instance type with 512 CPU units specified for that container, and that is the only task running on the container instance, that container could use the full 1,024 CPU unit share at any given time. However, if you launched another copy of the same task on that container instance, each task would be guaranteed a minimum of 512 CPU units when needed, and each container could float to higher CPU usage if the other container was not using it, but if both tasks were 100% active all of the time, they would be limited to 512 CPU units.

On Linux container instances, the Docker daemon on the container instance uses the CPU value to calculate the relative CPU share ratios for running containers. For more information, see CPU share constraint in the Docker documentation. The minimum valid CPU share value that the Linux kernel allows is 2. However, the CPU parameter is not required, and you can use CPU values below 2 in your container definitions. For CPU values below 2 (including null), the behavior varies based on your Amazon ECS container agent version:

Agent versions less than or equal to 1.1.0: Null and zero CPU values are passed to Docker as 0, which Docker then converts to 1,024 CPU shares. CPU values of 1 are passed to Docker as 1, which the Linux kernel converts to two CPU shares.

Agent versions greater than or equal to 1.2.0: Null, zero, and CPU values of 1 are passed to Docker as 2.

On Windows container instances, the CPU limit is enforced as an absolute limit, or a quota. Windows containers only have access to the specified amount of CPU that is described in the task definition. A null or zero CPU value is passed to Docker as 0, which Windows interprets as 1% of one CPU.

pub fn set_cpu(self, input: Option<i32>) -> Self

pub fn memory(self, input: i32) -> Self

The amount (in MiB) of memory to present to the container. If your container attempts to exceed the memory specified here, the container is killed. The total amount of memory reserved for all containers within a task must be lower than the task memory value, if one is specified. This parameter maps to Memory in the Create a container section of the Docker Remote API and the --memory option to docker run.

If using the Fargate launch type, this parameter is optional.

If using the EC2 launch type, you must specify either a task-level memory value or a container-level memory value. If you specify both a container-level memory and memoryReservation value, memory must be greater than memoryReservation. If you specify memoryReservation, then that value is subtracted from the available memory resources for the container instance on which the container is placed. Otherwise, the value of memory is used.

The Docker daemon reserves a minimum of 4 MiB of memory for a container, so you should not specify fewer than 4 MiB of memory for your containers.

pub fn set_memory(self, input: Option<i32>) -> Self

pub fn memory_reservation(self, input: i32) -> Self

The soft limit (in MiB) of memory to reserve for the container. When system memory is under heavy contention, Docker attempts to keep the container memory to this soft limit. However, your container can consume more memory when it needs to, up to either the hard limit specified with the memory parameter (if applicable), or all of the available memory on the container instance, whichever comes first. This parameter maps to MemoryReservation in the Create a container section of the Docker Remote API and the --memory-reservation option to docker run.

If a task-level memory value is not specified, you must specify a non-zero integer for one or both of memory or memoryReservation in a container definition. If you specify both, memory must be greater than memoryReservation. If you specify memoryReservation, then that value is subtracted from the available memory resources for the container instance on which the container is placed. Otherwise, the value of memory is used.

For example, if your container normally uses 128 MiB of memory, but occasionally bursts to 256 MiB of memory for short periods of time, you can set a memoryReservation of 128 MiB, and a memory hard limit of 300 MiB. This configuration would allow the container to only reserve 128 MiB of memory from the remaining resources on the container instance, but also allow the container to consume more memory resources when needed.

The Docker daemon reserves a minimum of 4 MiB of memory for a container, so you should not specify fewer than 4 MiB of memory for your containers.

pub fn set_memory_reservation(self, input: Option<i32>) -> Self

pub fn links(self, input: impl Into<String>) -> Self

pub fn set_links(self, input: Option<Vec<String>>) -> Self

pub fn port_mappings(self, input: impl Into<PortMapping>) -> Self

pub fn set_port_mappings(self, input: Option<Vec<PortMapping>>) -> Self

pub fn essential(self, input: bool) -> Self

If the essential parameter of a container is marked as true, and that container fails or stops for any reason, all other containers that are part of the task are stopped. If the essential parameter of a container is marked as false, then its failure does not affect the rest of the containers in a task. If this parameter is omitted, a container is assumed to be essential.

All tasks must have at least one essential container. If you have an application that is composed of multiple containers, you should group containers that are used for a common purpose into components, and separate the different components into multiple task definitions. For more information, see Application Architecture in the Amazon Elastic Container Service Developer Guide.

pub fn set_essential(self, input: Option<bool>) -> Self

pub fn entry_point(self, input: impl Into<String>) -> Self

pub fn set_entry_point(self, input: Option<Vec<String>>) -> Self

pub fn command(self, input: impl Into<String>) -> Self

pub fn set_command(self, input: Option<Vec<String>>) -> Self

pub fn environment(self, input: impl Into<KeyValuePair>) -> Self

pub fn set_environment(self, input: Option<Vec<KeyValuePair>>) -> Self

pub fn environment_files(self, input: impl Into<EnvironmentFile>) -> Self

pub fn set_environment_files(self, input: Option<Vec<EnvironmentFile>>) -> Self

pub fn mount_points(self, input: impl Into<MountPoint>) -> Self

pub fn set_mount_points(self, input: Option<Vec<MountPoint>>) -> Self

pub fn volumes_from(self, input: impl Into<VolumeFrom>) -> Self

pub fn set_volumes_from(self, input: Option<Vec<VolumeFrom>>) -> Self

pub fn linux_parameters(self, input: LinuxParameters) -> Self

Linux-specific modifications that are applied to the container, such as Linux kernel capabilities. For more information see KernelCapabilities.

This parameter is not supported for Windows containers.

pub fn set_linux_parameters(self, input: Option<LinuxParameters>) -> Self

pub fn secrets(self, input: impl Into<Secret>) -> Self

pub fn set_secrets(self, input: Option<Vec<Secret>>) -> Self

pub fn depends_on(self, input: impl Into<ContainerDependency>) -> Self

pub fn set_depends_on(self, input: Option<Vec<ContainerDependency>>) -> Self

pub fn start_timeout(self, input: i32) -> Self

Time duration (in seconds) to wait before giving up on resolving dependencies for a container. For example, you specify two containers in a task definition with containerA having a dependency on containerB reaching a COMPLETE, SUCCESS, or HEALTHY status. If a startTimeout value is specified for containerB and it does not reach the desired status within that time then containerA will give up and not start. This results in the task transitioning to a STOPPED state.

When the ECS_CONTAINER_START_TIMEOUT container agent configuration variable is used, it is enforced indendently from this start timeout value.

For tasks using the Fargate launch type, this parameter requires that the task or service uses platform version 1.3.0 or later.

For tasks using the EC2 launch type, your container instances require at least version 1.26.0 of the container agent to enable a container start timeout value. However, we recommend using the latest container agent version. For information about checking your agent version and updating to the latest version, see Updating the Amazon ECS Container Agent in the Amazon Elastic Container Service Developer Guide. If you are using an Amazon ECS-optimized Linux AMI, your instance needs at least version 1.26.0-1 of the ecs-init package. If your container instances are launched from version 20190301 or later, then they contain the required versions of the container agent and ecs-init. For more information, see Amazon ECS-optimized Linux AMI in the Amazon Elastic Container Service Developer Guide.

pub fn set_start_timeout(self, input: Option<i32>) -> Self

pub fn stop_timeout(self, input: i32) -> Self

Time duration (in seconds) to wait before the container is forcefully killed if it doesn't exit normally on its own.

For tasks using the Fargate launch type, the task or service requires platform version 1.3.0 or later. The max stop timeout value is 120 seconds and if the parameter is not specified, the default value of 30 seconds is used.

For tasks using the EC2 launch type, if the stopTimeout parameter is not specified, the value set for the Amazon ECS container agent configuration variable ECS_CONTAINER_STOP_TIMEOUT is used by default. If neither the stopTimeout parameter or the ECS_CONTAINER_STOP_TIMEOUT agent configuration variable are set, then the default values of 30 seconds for Linux containers and 30 seconds on Windows containers are used. Your container instances require at least version 1.26.0 of the container agent to enable a container stop timeout value. However, we recommend using the latest container agent version. For information about checking your agent version and updating to the latest version, see Updating the Amazon ECS Container Agent in the Amazon Elastic Container Service Developer Guide. If you are using an Amazon ECS-optimized Linux AMI, your instance needs at least version 1.26.0-1 of the ecs-init package. If your container instances are launched from version 20190301 or later, then they contain the required versions of the container agent and ecs-init. For more information, see Amazon ECS-optimized Linux AMI in the Amazon Elastic Container Service Developer Guide.

pub fn set_stop_timeout(self, input: Option<i32>) -> Self

pub fn hostname(self, input: impl Into<String>) -> Self

The hostname to use for your container. This parameter maps to Hostname in the Create a container section of the Docker Remote API and the --hostname option to docker run.

The hostname parameter is not supported if you are using the awsvpc network mode.

pub fn set_hostname(self, input: Option<String>) -> Self

pub fn user(self, input: impl Into<String>) -> Self

The user to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

When running tasks using the host network mode, you should not run containers using the root user (UID 0). It is considered best practice to use a non-root user.

You can specify the user using the following formats. If specifying a UID or GID, you must specify it as a positive integer.

user

user:group

uid

uid:gid

user:gid

uid:group

This parameter is not supported for Windows containers.

pub fn set_user(self, input: Option<String>) -> Self

pub fn working_directory(self, input: impl Into<String>) -> Self

The working directory in which to run commands inside the container. This parameter maps to WorkingDir in the Create a container section of the Docker Remote API and the --workdir option to docker run.

pub fn set_working_directory(self, input: Option<String>) -> Self

pub fn disable_networking(self, input: bool) -> Self

When this parameter is true, networking is disabled within the container. This parameter maps to NetworkDisabled in the Create a container section of the Docker Remote API.

This parameter is not supported for Windows containers.

pub fn set_disable_networking(self, input: Option<bool>) -> Self

pub fn privileged(self, input: bool) -> Self

When this parameter is true, the container is given elevated privileges on the host container instance (similar to the root user). This parameter maps to Privileged in the Create a container section of the Docker Remote API and the --privileged option to docker run.

This parameter is not supported for Windows containers or tasks run on AWS Fargate.

pub fn set_privileged(self, input: Option<bool>) -> Self

pub fn readonly_root_filesystem(self, input: bool) -> Self

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run.

This parameter is not supported for Windows containers.

pub fn set_readonly_root_filesystem(self, input: Option<bool>) -> Self

pub fn dns_servers(self, input: impl Into<String>) -> Self

pub fn set_dns_servers(self, input: Option<Vec<String>>) -> Self

pub fn dns_search_domains(self, input: impl Into<String>) -> Self

pub fn set_dns_search_domains(self, input: Option<Vec<String>>) -> Self

pub fn extra_hosts(self, input: impl Into<HostEntry>) -> Self

pub fn set_extra_hosts(self, input: Option<Vec<HostEntry>>) -> Self

pub fn docker_security_options(self, input: impl Into<String>) -> Self

pub fn set_docker_security_options(self, input: Option<Vec<String>>) -> Self

pub fn interactive(self, input: bool) -> Self

When this parameter is true, this allows you to deploy containerized applications that require stdin or a tty to be allocated. This parameter maps to OpenStdin in the Create a container section of the Docker Remote API and the --interactive option to docker run.

pub fn set_interactive(self, input: Option<bool>) -> Self

pub fn pseudo_terminal(self, input: bool) -> Self

When this parameter is true, a TTY is allocated. This parameter maps to Tty in the Create a container section of the Docker Remote API and the --tty option to docker run.

pub fn set_pseudo_terminal(self, input: Option<bool>) -> Self

pub fn docker_labels(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_docker_labels(self, input: Option<HashMap<String, String>>) -> Self

pub fn ulimits(self, input: impl Into<Ulimit>) -> Self

pub fn set_ulimits(self, input: Option<Vec<Ulimit>>) -> Self

pub fn log_configuration(self, input: LogConfiguration) -> Self

The log configuration specification for the container.

This parameter maps to LogConfig in the Create a container section of the Docker Remote API and the --log-driver option to docker run. By default, containers use the same logging driver that the Docker daemon uses. However the container may use a different logging driver than the Docker daemon by specifying a log driver with this parameter in the container definition. To use a different logging driver for a container, the log system must be configured properly on the container instance (or on a different log server for remote logging options). For more information on the options for different supported log drivers, see Configure logging drivers in the Docker documentation.

Amazon ECS currently supports a subset of the logging drivers available to the Docker daemon (shown in the LogConfiguration data type). Additional log drivers may be available in future releases of the Amazon ECS container agent.

This parameter requires version 1.18 of the Docker Remote API or greater on your container instance. To check the Docker Remote API version on your container instance, log in to your container instance and run the following command: sudo docker version --format '{{.Server.APIVersion}}'

The Amazon ECS container agent running on a container instance must register the logging drivers available on that instance with the ECS_AVAILABLE_LOGGING_DRIVERS environment variable before containers placed on that instance can use these log configuration options. For more information, see Amazon ECS Container Agent Configuration in the Amazon Elastic Container Service Developer Guide.

pub fn set_log_configuration(self, input: Option<LogConfiguration>) -> Self

pub fn health_check(self, input: HealthCheck) -> Self

The container health check command and associated configuration parameters for the container. This parameter maps to HealthCheck in the Create a container section of the Docker Remote API and the HEALTHCHECK parameter of docker run.

pub fn set_health_check(self, input: Option<HealthCheck>) -> Self

pub fn system_controls(self, input: impl Into<SystemControl>) -> Self

pub fn set_system_controls(self, input: Option<Vec<SystemControl>>) -> Self

pub fn resource_requirements(
self,
input: impl Into<ResourceRequirement>
) -> Self

pub fn set_resource_requirements(
self,
input: Option<Vec<ResourceRequirement>>
) -> Self

pub fn firelens_configuration(self, input: FirelensConfiguration) -> Self

The FireLens configuration for the container. This is used to specify and configure a log router for container logs. For more information, see Custom Log Routing in the Amazon Elastic Container Service Developer Guide.

pub fn set_firelens_configuration(
self,
input: Option<FirelensConfiguration>
) -> Self

pub fn build(self) -> ContainerDefinition

Consumes the builder and constructs a ContainerDefinition