Struct aws_sdk_ecs::model::container_instance::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ContainerInstance

Implementations

impl Builder

pub fn container_instance_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the container instance. The ARN contains the arn:aws:ecs namespace, followed by the Region of the container instance, the AWS account ID of the container instance owner, the container-instance namespace, and then the container instance ID. For example, arn:aws:ecs:region:aws_account_id:container-instance/container_instance_ID.

pub fn set_container_instance_arn(self, input: Option<String>) -> Self

pub fn ec2_instance_id(self, input: impl Into<String>) -> Self

The ID of the container instance. For Amazon EC2 instances, this value is the Amazon EC2 instance ID. For external instances, this value is the AWS Systems Manager managed instance ID.

pub fn set_ec2_instance_id(self, input: Option<String>) -> Self

pub fn capacity_provider_name(self, input: impl Into<String>) -> Self

The capacity provider associated with the container instance.

pub fn set_capacity_provider_name(self, input: Option<String>) -> Self

pub fn version(self, input: i64) -> Self

The version counter for the container instance. Every time a container instance experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS container instance state with CloudWatch Events, you can compare the version of a container instance reported by the Amazon ECS APIs with the version reported in CloudWatch Events for the container instance (inside the detail object) to verify that the version in your event stream is current.

pub fn set_version(self, input: Option<i64>) -> Self

pub fn version_info(self, input: VersionInfo) -> Self

The version information for the Amazon ECS container agent and Docker daemon running on the container instance.

pub fn set_version_info(self, input: Option<VersionInfo>) -> Self

pub fn remaining_resources(self, input: impl Into<Resource>) -> Self

pub fn set_remaining_resources(self, input: Option<Vec<Resource>>) -> Self

pub fn registered_resources(self, input: impl Into<Resource>) -> Self

pub fn set_registered_resources(self, input: Option<Vec<Resource>>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The status of the container instance. The valid values are REGISTERING, REGISTRATION_FAILED, ACTIVE, INACTIVE, DEREGISTERING, or DRAINING.

If your account has opted in to the awsvpcTrunking account setting, then any newly registered container instance will transition to a REGISTERING status while the trunk elastic network interface is provisioned for the instance. If the registration fails, the instance will transition to a REGISTRATION_FAILED status. You can describe the container instance and see the reason for failure in the statusReason parameter. Once the container instance is terminated, the instance transitions to a DEREGISTERING status while the trunk elastic network interface is deprovisioned. The instance then transitions to an INACTIVE status.

The ACTIVE status indicates that the container instance can accept tasks. The DRAINING indicates that new tasks are not placed on the container instance and any service tasks running on the container instance are removed if possible. For more information, see Container Instance Draining in the Amazon Elastic Container Service Developer Guide.

pub fn set_status(self, input: Option<String>) -> Self

pub fn status_reason(self, input: impl Into<String>) -> Self

The reason that the container instance reached its current status.

pub fn set_status_reason(self, input: Option<String>) -> Self

pub fn agent_connected(self, input: bool) -> Self

This parameter returns true if the agent is connected to Amazon ECS. Registered instances with an agent that may be unhealthy or stopped return false. Only instances connected to an agent can accept placement requests.

pub fn set_agent_connected(self, input: Option<bool>) -> Self

pub fn running_tasks_count(self, input: i32) -> Self

The number of tasks on the container instance that are in the RUNNING status.

pub fn set_running_tasks_count(self, input: Option<i32>) -> Self

pub fn pending_tasks_count(self, input: i32) -> Self

The number of tasks on the container instance that are in the PENDING status.

pub fn set_pending_tasks_count(self, input: Option<i32>) -> Self

pub fn agent_update_status(self, input: AgentUpdateStatus) -> Self

The status of the most recent agent update. If an update has never been requested, this value is NULL.

pub fn set_agent_update_status(self, input: Option<AgentUpdateStatus>) -> Self

pub fn attributes(self, input: impl Into<Attribute>) -> Self

pub fn set_attributes(self, input: Option<Vec<Attribute>>) -> Self

pub fn registered_at(self, input: Instant) -> Self

The Unix timestamp for when the container instance was registered.

pub fn set_registered_at(self, input: Option<Instant>) -> Self

pub fn attachments(self, input: impl Into<Attachment>) -> Self

pub fn set_attachments(self, input: Option<Vec<Attachment>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> ContainerInstance

Consumes the builder and constructs a ContainerInstance