Struct aws_sdk_ecs::model::Service

#[non_exhaustive]
pub struct Service {

    pub service_arn: Option<String>,
    pub service_name: Option<String>,
    pub cluster_arn: Option<String>,
    pub load_balancers: Option<Vec<LoadBalancer>>,
    pub service_registries: Option<Vec<ServiceRegistry>>,
    pub status: Option<String>,
    pub desired_count: i32,
    pub running_count: i32,
    pub pending_count: i32,
    pub launch_type: Option<LaunchType>,
    pub capacity_provider_strategy: Option<Vec<CapacityProviderStrategyItem>>,
    pub platform_version: Option<String>,
    pub task_definition: Option<String>,
    pub deployment_configuration: Option<DeploymentConfiguration>,
    pub task_sets: Option<Vec<TaskSet>>,
    pub deployments: Option<Vec<Deployment>>,
    pub role_arn: Option<String>,
    pub events: Option<Vec<ServiceEvent>>,
    pub created_at: Option<Instant>,
    pub placement_constraints: Option<Vec<PlacementConstraint>>,
    pub placement_strategy: Option<Vec<PlacementStrategy>>,
    pub network_configuration: Option<NetworkConfiguration>,
    pub health_check_grace_period_seconds: Option<i32>,
    pub scheduling_strategy: Option<SchedulingStrategy>,
    pub deployment_controller: Option<DeploymentController>,
    pub tags: Option<Vec<Tag>>,
    pub created_by: Option<String>,
    pub enable_ecs_managed_tags: bool,
    pub propagate_tags: Option<PropagateTags>,
    pub enable_execute_command: bool,

}

Details on a service within a cluster

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

service_arn: Option<String>

The ARN that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the Region of the service, the AWS account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

service_name: Option<String>

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

cluster_arn: Option<String>

The Amazon Resource Name (ARN) of the cluster that hosts the service.

load_balancers: Option<Vec<LoadBalancer>>

A list of Elastic Load Balancing load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

service_registries: Option<Vec<ServiceRegistry>>

The details of the service discovery registries to assign to this service. For more information, see Service Discovery.

status: Option<String>

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

desired_count: i32

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

running_count: i32

The number of tasks in the cluster that are in the RUNNING state.

pending_count: i32

The number of tasks in the cluster that are in the PENDING state.

launch_type: Option<LaunchType>

The infrastructure on which your service is running. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

capacity_provider_strategy: Option<Vec<CapacityProviderStrategyItem>>

The capacity provider strategy associated with the service.

platform_version: Option<String>

The platform version on which to run your service. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see AWS Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

task_definition: Option<String>

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

deployment_configuration: Option<DeploymentConfiguration>

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

task_sets: Option<Vec<TaskSet>>

Information about a set of Amazon ECS tasks in either an AWS CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

deployments: Option<Vec<Deployment>>

The current state of deployments for the service.

role_arn: Option<String>

The ARN of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

events: Option<Vec<ServiceEvent>>

The event stream for your service. A maximum of 100 of the latest events are displayed.

created_at: Option<Instant>

The Unix timestamp for when the service was created.

placement_constraints: Option<Vec<PlacementConstraint>>

The placement constraints for the tasks in the service.

placement_strategy: Option<Vec<PlacementStrategy>>

The placement strategy that determines how tasks for the service are placed.

network_configuration: Option<NetworkConfiguration>

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

health_check_grace_period_seconds: Option<i32>

The period of time, in seconds, that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks after a task has first started.

scheduling_strategy: Option<SchedulingStrategy>

The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions.

DAEMON-The daemon scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that you specify in your cluster. The service scheduler also evaluates the task placement constraints for running tasks and will stop tasks that do not meet the placement constraints.

Fargate tasks do not support the DAEMON scheduling strategy.

deployment_controller: Option<DeploymentController>

The deployment controller type the service is using. When using the DescribeServices API, this field is omitted if the service is using the ECS deployment controller type.

tags: Option<Vec<Tag>>

The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for AWS use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

created_by: Option<String>

The principal that created the service.

enable_ecs_managed_tags: bool

Specifies whether to enable Amazon ECS managed tags for the tasks in the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

propagate_tags: Option<PropagateTags>

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

enable_execute_command: bool

Whether or not the execute command functionality is enabled for the service. If true, the execute command functionality is enabled for all containers in tasks as part of the service.

Implementations

impl Service

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Service