Struct aws_sdk_ecs::model::Task

#[non_exhaustive]
pub struct Task {

    pub attachments: Option<Vec<Attachment>>,
    pub attributes: Option<Vec<Attribute>>,
    pub availability_zone: Option<String>,
    pub capacity_provider_name: Option<String>,
    pub cluster_arn: Option<String>,
    pub connectivity: Option<Connectivity>,
    pub connectivity_at: Option<Instant>,
    pub container_instance_arn: Option<String>,
    pub containers: Option<Vec<Container>>,
    pub cpu: Option<String>,
    pub created_at: Option<Instant>,
    pub desired_status: Option<String>,
    pub enable_execute_command: bool,
    pub execution_stopped_at: Option<Instant>,
    pub group: Option<String>,
    pub health_status: Option<HealthStatus>,
    pub inference_accelerators: Option<Vec<InferenceAccelerator>>,
    pub last_status: Option<String>,
    pub launch_type: Option<LaunchType>,
    pub memory: Option<String>,
    pub overrides: Option<TaskOverride>,
    pub platform_version: Option<String>,
    pub pull_started_at: Option<Instant>,
    pub pull_stopped_at: Option<Instant>,
    pub started_at: Option<Instant>,
    pub started_by: Option<String>,
    pub stop_code: Option<TaskStopCode>,
    pub stopped_at: Option<Instant>,
    pub stopped_reason: Option<String>,
    pub stopping_at: Option<Instant>,
    pub tags: Option<Vec<Tag>>,
    pub task_arn: Option<String>,
    pub task_definition_arn: Option<String>,
    pub version: i64,
    pub ephemeral_storage: Option<EphemeralStorage>,

}

Details on a task in a cluster.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

attachments: Option<Vec<Attachment>>

The Elastic Network Adapter associated with the task if the task uses the awsvpc network mode.

attributes: Option<Vec<Attribute>>

The attributes of the task

availability_zone: Option<String>

The availability zone of the task.

capacity_provider_name: Option<String>

The capacity provider associated with the task.

cluster_arn: Option<String>

The ARN of the cluster that hosts the task.

connectivity: Option<Connectivity>

The connectivity status of a task.

connectivity_at: Option<Instant>

The Unix timestamp for when the task last went into CONNECTED status.

container_instance_arn: Option<String>

The ARN of the container instances that host the task.

containers: Option<Vec<Container>>

The containers associated with the task.

cpu: Option<String>

The number of CPU units used by the task as expressed in a task definition. It can be expressed as an integer using CPU units, for example 1024. It can also be expressed as a string using vCPUs, for example 1 vCPU or 1 vcpu. String values are converted to an integer indicating the CPU units when the task definition is registered.

If you are using the EC2 launch type, this field is optional. Supported values are between 128 CPU units (0.125 vCPUs) and 10240 CPU units (10 vCPUs).

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the memory parameter:

256 (.25 vCPU) - Available memory values: 512 (0.5 GB), 1024 (1 GB), 2048 (2 GB)

512 (.5 vCPU) - Available memory values: 1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB)

1024 (1 vCPU) - Available memory values: 2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB)

2048 (2 vCPU) - Available memory values: Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB)

4096 (4 vCPU) - Available memory values: Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB)

created_at: Option<Instant>

The Unix timestamp for when the task was created (the task entered the PENDING state).

desired_status: Option<String>

The desired status of the task. For more information, see Task Lifecycle.

enable_execute_command: bool

Whether or not execute command functionality is enabled for this task. If true, this enables execute command functionality on all containers in the task.

execution_stopped_at: Option<Instant>

The Unix timestamp for when the task execution stopped.

group: Option<String>

The name of the task group associated with the task.

health_status: Option<HealthStatus>

The health status for the task, which is determined by the health of the essential containers in the task. If all essential containers in the task are reporting as HEALTHY, then the task status also reports as HEALTHY. If any essential containers in the task are reporting as UNHEALTHY or UNKNOWN, then the task status also reports as UNHEALTHY or UNKNOWN, accordingly.

The Amazon ECS container agent does not monitor or report on Docker health checks that are embedded in a container image (such as those specified in a parent image or from the image's Dockerfile) and not specified in the container definition. Health check parameters that are specified in a container definition override any Docker health checks that exist in the container image.

inference_accelerators: Option<Vec<InferenceAccelerator>>

The Elastic Inference accelerator associated with the task.

last_status: Option<String>

The last known status of the task. For more information, see Task Lifecycle.

launch_type: Option<LaunchType>

The infrastructure on which your task is running. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

memory: Option<String>

The amount of memory (in MiB) used by the task as expressed in a task definition. It can be expressed as an integer using MiB, for example 1024. It can also be expressed as a string using GB, for example 1GB or 1 GB. String values are converted to an integer indicating the MiB when the task definition is registered.

If you are using the EC2 launch type, this field is optional.

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the cpu parameter:

512 (0.5 GB), 1024 (1 GB), 2048 (2 GB) - Available cpu values: 256 (.25 vCPU)

1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB) - Available cpu values: 512 (.5 vCPU)

2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB) - Available cpu values: 1024 (1 vCPU)

Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB) - Available cpu values: 2048 (2 vCPU)

Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB) - Available cpu values: 4096 (4 vCPU)

overrides: Option<TaskOverride>

One or more container overrides.

platform_version: Option<String>

The platform version on which your task is running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see AWS Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pull_started_at: Option<Instant>

The Unix timestamp for when the container image pull began.

pull_stopped_at: Option<Instant>

The Unix timestamp for when the container image pull completed.

started_at: Option<Instant>

The Unix timestamp for when the task started (the task transitioned from the PENDING state to the RUNNING state).

started_by: Option<String>

The tag specified when a task is started. If the task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

stop_code: Option<TaskStopCode>

The stop code indicating why a task was stopped. The stoppedReason may contain additional details.

stopped_at: Option<Instant>

The Unix timestamp for when the task was stopped (the task transitioned from the RUNNING state to the STOPPED state).

stopped_reason: Option<String>

The reason that the task was stopped.

stopping_at: Option<Instant>

The Unix timestamp for when the task stops (transitions from the RUNNING state to STOPPED).

tags: Option<Vec<Tag>>

The metadata that you apply to the task to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for AWS use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

task_arn: Option<String>

The Amazon Resource Name (ARN) of the task.

task_definition_arn: Option<String>

The ARN of the task definition that creates the task.

version: i64

The version counter for the task. Every time a task experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS task state with CloudWatch Events, you can compare the version of a task reported by the Amazon ECS API actions with the version reported in CloudWatch Events for the task (inside the detail object) to verify that the version in your event stream is current.

ephemeral_storage: Option<EphemeralStorage>

The ephemeral storage settings for the task.

Implementations

impl Task

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Task