Struct aws_sdk_ecs::model::TaskSet

#[non_exhaustive]
pub struct TaskSet {

    pub id: Option<String>,
    pub task_set_arn: Option<String>,
    pub service_arn: Option<String>,
    pub cluster_arn: Option<String>,
    pub started_by: Option<String>,
    pub external_id: Option<String>,
    pub status: Option<String>,
    pub task_definition: Option<String>,
    pub computed_desired_count: i32,
    pub pending_count: i32,
    pub running_count: i32,
    pub created_at: Option<Instant>,
    pub updated_at: Option<Instant>,
    pub launch_type: Option<LaunchType>,
    pub capacity_provider_strategy: Option<Vec<CapacityProviderStrategyItem>>,
    pub platform_version: Option<String>,
    pub network_configuration: Option<NetworkConfiguration>,
    pub load_balancers: Option<Vec<LoadBalancer>>,
    pub service_registries: Option<Vec<ServiceRegistry>>,
    pub scale: Option<Scale>,
    pub stability_status: Option<StabilityStatus>,
    pub stability_status_at: Option<Instant>,
    pub tags: Option<Vec<Tag>>,

}

Information about a set of Amazon ECS tasks in either an AWS CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: Option<String>

The ID of the task set.

task_set_arn: Option<String>

The Amazon Resource Name (ARN) of the task set.

service_arn: Option<String>

The Amazon Resource Name (ARN) of the service the task set exists in.

cluster_arn: Option<String>

The Amazon Resource Name (ARN) of the cluster that the service that hosts the task set exists in.

started_by: Option<String>

The tag specified when a task set is started. If the task set is created by an AWS CodeDeploy deployment, the startedBy parameter is CODE_DEPLOY. For a task set created for an external deployment, the startedBy field isn't used.

external_id: Option<String>

The external ID associated with the task set.

If a task set is created by an AWS CodeDeploy deployment, the externalId parameter contains the AWS CodeDeploy deployment ID.

If a task set is created for an external deployment and is associated with a service discovery registry, the externalId parameter contains the ECS_TASK_SET_EXTERNAL_ID AWS Cloud Map attribute.

status: Option<String>

The status of the task set. The following describes each state:

PRIMARY

The task set is serving production traffic.

ACTIVE

The task set is not serving production traffic.

DRAINING

The tasks in the task set are being stopped and their corresponding targets are being deregistered from their target group.

task_definition: Option<String>

The task definition the task set is using.

computed_desired_count: i32

The computed desired count for the task set. This is calculated by multiplying the service's desiredCount by the task set's scale percentage. The result is always rounded up. For example, if the computed desired count is 1.2, it rounds up to 2 tasks.

pending_count: i32

The number of tasks in the task set that are in the PENDING status during a deployment. A task in the PENDING state is preparing to enter the RUNNING state. A task set enters the PENDING status when it launches for the first time or when it is restarted after being in the STOPPED state.

running_count: i32

The number of tasks in the task set that are in the RUNNING status during a deployment. A task in the RUNNING state is running and ready for use.

created_at: Option<Instant>

The Unix timestamp for when the task set was created.

updated_at: Option<Instant>

The Unix timestamp for when the task set was last updated.

launch_type: Option<LaunchType>

The launch type the tasks in the task set are using. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

capacity_provider_strategy: Option<Vec<CapacityProviderStrategyItem>>

The capacity provider strategy associated with the task set.

platform_version: Option<String>

The AWS Fargate platform version on which the tasks in the task set are running. A platform version is only specified for tasks run on AWS Fargate. For more information, see AWS Fargate platform versions in the Amazon Elastic Container Service Developer Guide.

network_configuration: Option<NetworkConfiguration>

The network configuration for the task set.

load_balancers: Option<Vec<LoadBalancer>>

Details on a load balancer that is used with a task set.

service_registries: Option<Vec<ServiceRegistry>>

The details of the service discovery registries to assign to this task set. For more information, see Service discovery.

scale: Option<Scale>

A floating-point percentage of the desired number of tasks to place and keep running in the task set.

stability_status: Option<StabilityStatus>

The stability status, which indicates whether the task set has reached a steady state. If the following conditions are met, the task set will be in STEADY_STATE:

The task runningCount is equal to the computedDesiredCount.

The pendingCount is 0.

There are no tasks running on container instances in the DRAINING status.

All tasks are reporting a healthy status from the load balancers, service discovery, and container health checks.

If any of those conditions are not met, the stability status returns STABILIZING.

stability_status_at: Option<Instant>

The Unix timestamp for when the task set stability status was retrieved.

tags: Option<Vec<Tag>>

The metadata that you apply to the task set to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for AWS use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

Implementations

impl TaskSet

pub fn builder() -> Builder

Creates a new builder-style object to manufacture TaskSet