Module aws_sdk_ecs::error

Modules

access_denied_error
See AccessDeniedError

attribute_limit_exceeded_error
See AttributeLimitExceededError

blocked_error
See BlockedError

client_error
See ClientError

cluster_contains_container_instances_error
See ClusterContainsContainerInstancesError

cluster_contains_services_error
See ClusterContainsServicesError

cluster_contains_tasks_error
See ClusterContainsTasksError

cluster_not_found_error
See ClusterNotFoundError

invalid_parameter_error
See InvalidParameterError

limit_exceeded_error
See LimitExceededError

missing_version_error
See MissingVersionError

no_update_available_error
See NoUpdateAvailableError

platform_task_definition_incompatibility_error
See PlatformTaskDefinitionIncompatibilityError

platform_unknown_error
See PlatformUnknownError

resource_in_use_error
See ResourceInUseError

resource_not_found_error
See ResourceNotFoundError

server_error
See ServerError

service_not_active_error
See ServiceNotActiveError

service_not_found_error
See ServiceNotFoundError

target_not_connected_error
See TargetNotConnectedError

target_not_found_error
See TargetNotFoundError

task_set_not_found_error
See TaskSetNotFoundError

unsupported_feature_error
See UnsupportedFeatureError

update_in_progress_error
See UpdateInProgressError

Structs

AccessDeniedError
You do not have authorization to perform the requested action.

AttributeLimitExceededError
You can apply up to 10 custom attributes per resource. You can view the attributes of

BlockedError
Your AWS account has been blocked. For more information, contact AWS Support.

ClientError
These errors are usually caused by a client action, such as using an action or

ClusterContainsContainerInstancesError
You cannot delete a cluster that has registered container instances. First, deregister

ClusterContainsServicesError
You cannot delete a cluster that contains services. First, update the service to

ClusterContainsTasksError
You cannot delete a cluster that has active tasks.

ClusterNotFoundError
The specified cluster could not be found. You can view your available clusters with

CreateCapacityProviderError

CreateClusterError

CreateServiceError

CreateTaskSetError

DeleteAccountSettingError

DeleteAttributesError

DeleteCapacityProviderError

DeleteClusterError

DeleteServiceError

DeleteTaskSetError

DeregisterContainerInstanceError

DeregisterTaskDefinitionError

DescribeCapacityProvidersError

DescribeClustersError

DescribeContainerInstancesError

DescribeServicesError

DescribeTaskDefinitionError

DescribeTaskSetsError

DescribeTasksError

DiscoverPollEndpointError

ExecuteCommandError

InvalidParameterError
The specified parameter is invalid. Review the available parameters for the API

LimitExceededError
The limit for the resource has been exceeded.

ListAccountSettingsError

ListAttributesError

ListClustersError

ListContainerInstancesError

ListServicesError

ListTagsForResourceError

ListTaskDefinitionFamiliesError

ListTaskDefinitionsError

ListTasksError

MissingVersionError
Amazon ECS is unable to determine the current version of the Amazon ECS container agent on the

NoUpdateAvailableError
There is no update available for this Amazon ECS container agent. This could be because the

PlatformTaskDefinitionIncompatibilityError
The specified platform version does not satisfy the task definition's required

PlatformUnknownError
The specified platform version does not exist.

PutAccountSettingDefaultError

PutAccountSettingError

PutAttributesError

PutClusterCapacityProvidersError

RegisterContainerInstanceError

RegisterTaskDefinitionError

ResourceInUseError
The specified resource is in-use and cannot be removed.

ResourceNotFoundError
The specified resource could not be found.

RunTaskError

ServerError
These errors are usually caused by a server issue.

ServiceNotActiveError
The specified service is not active. You can't update a service that is inactive. If

ServiceNotFoundError
The specified service could not be found. You can view your available services with

StartTaskError

StopTaskError

SubmitAttachmentStateChangesError

SubmitContainerStateChangeError

SubmitTaskStateChangeError

TagResourceError

TargetNotConnectedError
The target container is not properly configured with the execute command agent or the

TargetNotFoundError
The specified target could not be found. You can view your available container

TaskSetNotFoundError
The specified task set could not be found. You can view your available task sets with

UnsupportedFeatureError
The specified task is not supported in this Region.

UntagResourceError

UpdateCapacityProviderError

UpdateClusterError

UpdateClusterSettingsError

UpdateContainerAgentError

UpdateContainerInstancesStateError

UpdateInProgressError
There is already a current Amazon ECS container agent update in progress on the specified

UpdateServiceError

UpdateServicePrimaryTaskSetError

UpdateTaskSetError

Enums

CreateCapacityProviderErrorKind

CreateClusterErrorKind

CreateServiceErrorKind

CreateTaskSetErrorKind

DeleteAccountSettingErrorKind

DeleteAttributesErrorKind

DeleteCapacityProviderErrorKind

DeleteClusterErrorKind

DeleteServiceErrorKind

DeleteTaskSetErrorKind

DeregisterContainerInstanceErrorKind

DeregisterTaskDefinitionErrorKind

DescribeCapacityProvidersErrorKind

DescribeClustersErrorKind

DescribeContainerInstancesErrorKind

DescribeServicesErrorKind

DescribeTaskDefinitionErrorKind

DescribeTaskSetsErrorKind

DescribeTasksErrorKind

DiscoverPollEndpointErrorKind

ExecuteCommandErrorKind

ListAccountSettingsErrorKind

ListAttributesErrorKind

ListClustersErrorKind

ListContainerInstancesErrorKind

ListServicesErrorKind

ListTagsForResourceErrorKind

ListTaskDefinitionFamiliesErrorKind

ListTaskDefinitionsErrorKind

ListTasksErrorKind

PutAccountSettingDefaultErrorKind

PutAccountSettingErrorKind

PutAttributesErrorKind

PutClusterCapacityProvidersErrorKind

RegisterContainerInstanceErrorKind

RegisterTaskDefinitionErrorKind

RunTaskErrorKind

StartTaskErrorKind

StopTaskErrorKind

SubmitAttachmentStateChangesErrorKind

SubmitContainerStateChangeErrorKind

SubmitTaskStateChangeErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateCapacityProviderErrorKind

UpdateClusterErrorKind

UpdateClusterSettingsErrorKind

UpdateContainerAgentErrorKind

UpdateContainerInstancesStateErrorKind

UpdateServiceErrorKind

UpdateServicePrimaryTaskSetErrorKind

UpdateTaskSetErrorKind