Struct aws_sdk_ecs::model::service::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Service

Implementations

impl Builder

pub fn service_arn(self, input: impl Into<String>) -> Self

The ARN that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the Region of the service, the Amazon Web Services account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

pub fn set_service_arn(self, input: Option<String>) -> Self

The ARN that identifies the service. The ARN contains the arn:aws:ecs namespace, followed by the Region of the service, the Amazon Web Services account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

pub fn service_name(self, input: impl Into<String>) -> Self

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

pub fn set_service_name(self, input: Option<String>) -> Self

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, underscores, and hyphens are allowed. Service names must be unique within a cluster, but you can have similarly named services in multiple clusters within a Region or across multiple Regions.

pub fn cluster_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the cluster that hosts the service.

pub fn set_cluster_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) of the cluster that hosts the service.

pub fn load_balancers(self, input: impl Into<LoadBalancer>) -> Self

Appends an item to load_balancers.

To override the contents of this collection use set_load_balancers.

A list of Elastic Load Balancing load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

pub fn set_load_balancers(self, input: Option<Vec<LoadBalancer>>) -> Self

A list of Elastic Load Balancing load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

pub fn service_registries(self, input: impl Into<ServiceRegistry>) -> Self

Appends an item to service_registries.

To override the contents of this collection use set_service_registries.

The details of the service discovery registries to assign to this service. For more information, see Service Discovery.

pub fn set_service_registries(self, input: Option<Vec<ServiceRegistry>>) -> Self

The details of the service discovery registries to assign to this service. For more information, see Service Discovery.

pub fn status(self, input: impl Into<String>) -> Self

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

pub fn set_status(self, input: Option<String>) -> Self

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

pub fn desired_count(self, input: i32) -> Self

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

pub fn set_desired_count(self, input: Option<i32>) -> Self

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

pub fn running_count(self, input: i32) -> Self

The number of tasks in the cluster that are in the RUNNING state.

pub fn set_running_count(self, input: Option<i32>) -> Self

The number of tasks in the cluster that are in the RUNNING state.

pub fn pending_count(self, input: i32) -> Self

The number of tasks in the cluster that are in the PENDING state.

pub fn set_pending_count(self, input: Option<i32>) -> Self

The number of tasks in the cluster that are in the PENDING state.

pub fn launch_type(self, input: LaunchType) -> Self

The launch type the service is using. When using the DescribeServices API, this field is omitted if the service was created using a capacity provider strategy.

pub fn set_launch_type(self, input: Option<LaunchType>) -> Self

The launch type the service is using. When using the DescribeServices API, this field is omitted if the service was created using a capacity provider strategy.

pub fn capacity_provider_strategy(
self,
input: impl Into<CapacityProviderStrategyItem>
) -> Self

Appends an item to capacity_provider_strategy.

To override the contents of this collection use set_capacity_provider_strategy.

The capacity provider strategy the service is using. When using the DescribeServices API, this field is omitted if the service was created using a launch type.

pub fn set_capacity_provider_strategy(
self,
input: Option<Vec<CapacityProviderStrategyItem>>
) -> Self

The capacity provider strategy the service is using. When using the DescribeServices API, this field is omitted if the service was created using a launch type.

pub fn platform_version(self, input: impl Into<String>) -> Self

The platform version on which to run your service. A platform version is only specified for tasks hosted on Fargate. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pub fn set_platform_version(self, input: Option<String>) -> Self

The platform version on which to run your service. A platform version is only specified for tasks hosted on Fargate. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pub fn platform_family(self, input: impl Into<String>) -> Self

The operating system that your tasks in the service are running on. A platform family is specified only for tasks using the Fargate launch type.

All tasks that run as part of this service must use the same platformFamily value as the service, for example, LINUX.

pub fn set_platform_family(self, input: Option<String>) -> Self

The operating system that your tasks in the service are running on. A platform family is specified only for tasks using the Fargate launch type.

All tasks that run as part of this service must use the same platformFamily value as the service, for example, LINUX.

pub fn task_definition(self, input: impl Into<String>) -> Self

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

pub fn set_task_definition(self, input: Option<String>) -> Self

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

pub fn deployment_configuration(self, input: DeploymentConfiguration) -> Self

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

pub fn set_deployment_configuration(
self,
input: Option<DeploymentConfiguration>
) -> Self

Optional deployment parameters that control how many tasks run during the deployment and the ordering of stopping and starting tasks.

pub fn task_sets(self, input: impl Into<TaskSet>) -> Self

Appends an item to task_sets.

To override the contents of this collection use set_task_sets.

Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

pub fn set_task_sets(self, input: Option<Vec<TaskSet>>) -> Self

Information about a set of Amazon ECS tasks in either an CodeDeploy or an EXTERNAL deployment. An Amazon ECS task set includes details such as the desired number of tasks, how many tasks are running, and whether the task set serves production traffic.

pub fn deployments(self, input: impl Into<Deployment>) -> Self

Appends an item to deployments.

To override the contents of this collection use set_deployments.

The current state of deployments for the service.

pub fn set_deployments(self, input: Option<Vec<Deployment>>) -> Self

The current state of deployments for the service.

pub fn role_arn(self, input: impl Into<String>) -> Self

The ARN of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

pub fn set_role_arn(self, input: Option<String>) -> Self

The ARN of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

pub fn events(self, input: impl Into<ServiceEvent>) -> Self

Appends an item to events.

To override the contents of this collection use set_events.

The event stream for your service. A maximum of 100 of the latest events are displayed.

pub fn set_events(self, input: Option<Vec<ServiceEvent>>) -> Self

The event stream for your service. A maximum of 100 of the latest events are displayed.

pub fn created_at(self, input: Instant) -> Self

The Unix timestamp for when the service was created.

pub fn set_created_at(self, input: Option<Instant>) -> Self

The Unix timestamp for when the service was created.

pub fn placement_constraints(
self,
input: impl Into<PlacementConstraint>
) -> Self

Appends an item to placement_constraints.

To override the contents of this collection use set_placement_constraints.

The placement constraints for the tasks in the service.

pub fn set_placement_constraints(
self,
input: Option<Vec<PlacementConstraint>>
) -> Self

The placement constraints for the tasks in the service.

pub fn placement_strategy(self, input: impl Into<PlacementStrategy>) -> Self

Appends an item to placement_strategy.

To override the contents of this collection use set_placement_strategy.

The placement strategy that determines how tasks for the service are placed.

pub fn set_placement_strategy(
self,
input: Option<Vec<PlacementStrategy>>
) -> Self

The placement strategy that determines how tasks for the service are placed.

pub fn network_configuration(self, input: NetworkConfiguration) -> Self

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

The VPC subnet and security group configuration for tasks that receive their own elastic network interface by using the awsvpc networking mode.

pub fn health_check_grace_period_seconds(self, input: i32) -> Self

The period of time, in seconds, that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks after a task has first started.

pub fn set_health_check_grace_period_seconds(self, input: Option<i32>) -> Self

The period of time, in seconds, that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks after a task has first started.

pub fn scheduling_strategy(self, input: SchedulingStrategy) -> Self

The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions.

DAEMON-The daemon scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that you specify in your cluster. The service scheduler also evaluates the task placement constraints for running tasks and will stop tasks that do not meet the placement constraints.

Fargate tasks do not support the DAEMON scheduling strategy.

pub fn set_scheduling_strategy(self, input: Option<SchedulingStrategy>) -> Self

The scheduling strategy to use for the service. For more information, see Services.

There are two service scheduler strategies available:

REPLICA-The replica scheduling strategy places and maintains the desired number of tasks across your cluster. By default, the service scheduler spreads tasks across Availability Zones. You can use task placement strategies and constraints to customize task placement decisions.

DAEMON-The daemon scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that you specify in your cluster. The service scheduler also evaluates the task placement constraints for running tasks and will stop tasks that do not meet the placement constraints.

Fargate tasks do not support the DAEMON scheduling strategy.

pub fn deployment_controller(self, input: DeploymentController) -> Self

The deployment controller type the service is using. When using the DescribeServices API, this field is omitted if the service is using the ECS deployment controller type.

pub fn set_deployment_controller(
self,
input: Option<DeploymentController>
) -> Self

The deployment controller type the service is using. When using the DescribeServices API, this field is omitted if the service is using the ECS deployment controller type.

pub fn tags(self, input: impl Into<Tag>) -> Self

Appends an item to tags.

To override the contents of this collection use set_tags.

The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

The metadata that you apply to the service to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define.

The following basic restrictions apply to tags:

Maximum number of tags per resource - 50

For each resource, each tag key must be unique, and each tag key can have only one value.

Maximum key length - 128 Unicode characters in UTF-8

Maximum value length - 256 Unicode characters in UTF-8

If your tagging schema is used across multiple services and resources, remember that other services may have restrictions on allowed characters. Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following characters: + - = . _ : / @.

Tag keys and values are case-sensitive.

Do not use aws:, AWS:, or any upper or lowercase combination of such as a prefix for either keys or values as it is reserved for Amazon Web Services use. You cannot edit or delete tag keys or values with this prefix. Tags with this prefix do not count against your tags per resource limit.

pub fn created_by(self, input: impl Into<String>) -> Self

The principal that created the service.

pub fn set_created_by(self, input: Option<String>) -> Self

The principal that created the service.

pub fn enable_ecs_managed_tags(self, input: bool) -> Self

Specifies whether to enable Amazon ECS managed tags for the tasks in the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

pub fn set_enable_ecs_managed_tags(self, input: Option<bool>) -> Self

Specifies whether to enable Amazon ECS managed tags for the tasks in the service. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

pub fn propagate_tags(self, input: PropagateTags) -> Self

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

pub fn set_propagate_tags(self, input: Option<PropagateTags>) -> Self

Specifies whether to propagate the tags from the task definition or the service to the task. If no value is specified, the tags are not propagated.

pub fn enable_execute_command(self, input: bool) -> Self

Whether or not the execute command functionality is enabled for the service. If true, the execute command functionality is enabled for all containers in tasks as part of the service.

pub fn set_enable_execute_command(self, input: Option<bool>) -> Self

Whether or not the execute command functionality is enabled for the service. If true, the execute command functionality is enabled for all containers in tasks as part of the service.

pub fn build(self) -> Service

Consumes the builder and constructs a Service