Struct aws_sdk_ecs::model::task::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Task

Implementations

impl Builder

pub fn attachments(self, input: impl Into<Attachment>) -> Self

pub fn set_attachments(self, input: Option<Vec<Attachment>>) -> Self

pub fn attributes(self, input: impl Into<Attribute>) -> Self

pub fn set_attributes(self, input: Option<Vec<Attribute>>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The availability zone of the task.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn capacity_provider_name(self, input: impl Into<String>) -> Self

The capacity provider associated with the task.

pub fn set_capacity_provider_name(self, input: Option<String>) -> Self

pub fn cluster_arn(self, input: impl Into<String>) -> Self

The ARN of the cluster that hosts the task.

pub fn set_cluster_arn(self, input: Option<String>) -> Self

pub fn connectivity(self, input: Connectivity) -> Self

The connectivity status of a task.

pub fn set_connectivity(self, input: Option<Connectivity>) -> Self

pub fn connectivity_at(self, input: Instant) -> Self

The Unix timestamp for when the task last went into CONNECTED status.

pub fn set_connectivity_at(self, input: Option<Instant>) -> Self

pub fn container_instance_arn(self, input: impl Into<String>) -> Self

The ARN of the container instances that host the task.

pub fn set_container_instance_arn(self, input: Option<String>) -> Self

pub fn containers(self, input: impl Into<Container>) -> Self

pub fn set_containers(self, input: Option<Vec<Container>>) -> Self

pub fn cpu(self, input: impl Into<String>) -> Self

The number of CPU units used by the task as expressed in a task definition. It can be expressed as an integer using CPU units, for example 1024. It can also be expressed as a string using vCPUs, for example 1 vCPU or 1 vcpu. String values are converted to an integer indicating the CPU units when the task definition is registered.

If you are using the EC2 launch type, this field is optional. Supported values are between 128 CPU units (0.125 vCPUs) and 10240 CPU units (10 vCPUs).

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the memory parameter:

256 (.25 vCPU) - Available memory values: 512 (0.5 GB), 1024 (1 GB), 2048 (2 GB)

512 (.5 vCPU) - Available memory values: 1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB)

1024 (1 vCPU) - Available memory values: 2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB)

2048 (2 vCPU) - Available memory values: Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB)

4096 (4 vCPU) - Available memory values: Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB)

pub fn set_cpu(self, input: Option<String>) -> Self

pub fn created_at(self, input: Instant) -> Self

The Unix timestamp for when the task was created (the task entered the PENDING state).

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn desired_status(self, input: impl Into<String>) -> Self

The desired status of the task. For more information, see Task Lifecycle.

pub fn set_desired_status(self, input: Option<String>) -> Self

pub fn enable_execute_command(self, input: bool) -> Self

Whether or not execute command functionality is enabled for this task. If true, this enables execute command functionality on all containers in the task.

pub fn set_enable_execute_command(self, input: Option<bool>) -> Self

pub fn execution_stopped_at(self, input: Instant) -> Self

The Unix timestamp for when the task execution stopped.

pub fn set_execution_stopped_at(self, input: Option<Instant>) -> Self

pub fn group(self, input: impl Into<String>) -> Self

The name of the task group associated with the task.

pub fn set_group(self, input: Option<String>) -> Self

pub fn health_status(self, input: HealthStatus) -> Self

The health status for the task, which is determined by the health of the essential containers in the task. If all essential containers in the task are reporting as HEALTHY, then the task status also reports as HEALTHY. If any essential containers in the task are reporting as UNHEALTHY or UNKNOWN, then the task status also reports as UNHEALTHY or UNKNOWN, accordingly.

The Amazon ECS container agent does not monitor or report on Docker health checks that are embedded in a container image (such as those specified in a parent image or from the image's Dockerfile) and not specified in the container definition. Health check parameters that are specified in a container definition override any Docker health checks that exist in the container image.

pub fn set_health_status(self, input: Option<HealthStatus>) -> Self

pub fn inference_accelerators(
self,
input: impl Into<InferenceAccelerator>
) -> Self

pub fn set_inference_accelerators(
self,
input: Option<Vec<InferenceAccelerator>>
) -> Self

pub fn last_status(self, input: impl Into<String>) -> Self

The last known status of the task. For more information, see Task Lifecycle.

pub fn set_last_status(self, input: Option<String>) -> Self

pub fn launch_type(self, input: LaunchType) -> Self

The infrastructure on which your task is running. For more information, see Amazon ECS launch types in the Amazon Elastic Container Service Developer Guide.

pub fn set_launch_type(self, input: Option<LaunchType>) -> Self

pub fn memory(self, input: impl Into<String>) -> Self

The amount of memory (in MiB) used by the task as expressed in a task definition. It can be expressed as an integer using MiB, for example 1024. It can also be expressed as a string using GB, for example 1GB or 1 GB. String values are converted to an integer indicating the MiB when the task definition is registered.

If you are using the EC2 launch type, this field is optional.

If you are using the Fargate launch type, this field is required and you must use one of the following values, which determines your range of supported values for the cpu parameter:

512 (0.5 GB), 1024 (1 GB), 2048 (2 GB) - Available cpu values: 256 (.25 vCPU)

1024 (1 GB), 2048 (2 GB), 3072 (3 GB), 4096 (4 GB) - Available cpu values: 512 (.5 vCPU)

2048 (2 GB), 3072 (3 GB), 4096 (4 GB), 5120 (5 GB), 6144 (6 GB), 7168 (7 GB), 8192 (8 GB) - Available cpu values: 1024 (1 vCPU)

Between 4096 (4 GB) and 16384 (16 GB) in increments of 1024 (1 GB) - Available cpu values: 2048 (2 vCPU)

Between 8192 (8 GB) and 30720 (30 GB) in increments of 1024 (1 GB) - Available cpu values: 4096 (4 vCPU)

pub fn set_memory(self, input: Option<String>) -> Self

pub fn overrides(self, input: TaskOverride) -> Self

One or more container overrides.

pub fn set_overrides(self, input: Option<TaskOverride>) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

The platform version on which your task is running. A platform version is only specified for tasks using the Fargate launch type. If one is not specified, the LATEST platform version is used by default. For more information, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn pull_started_at(self, input: Instant) -> Self

The Unix timestamp for when the container image pull began.

pub fn set_pull_started_at(self, input: Option<Instant>) -> Self

pub fn pull_stopped_at(self, input: Instant) -> Self

The Unix timestamp for when the container image pull completed.

pub fn set_pull_stopped_at(self, input: Option<Instant>) -> Self

pub fn started_at(self, input: Instant) -> Self

The Unix timestamp for when the task started (the task transitioned from the PENDING state to the RUNNING state).

pub fn set_started_at(self, input: Option<Instant>) -> Self

pub fn started_by(self, input: impl Into<String>) -> Self

The tag specified when a task is started. If the task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

pub fn set_started_by(self, input: Option<String>) -> Self

pub fn stop_code(self, input: TaskStopCode) -> Self

The stop code indicating why a task was stopped. The stoppedReason may contain additional details.

pub fn set_stop_code(self, input: Option<TaskStopCode>) -> Self

pub fn stopped_at(self, input: Instant) -> Self

The Unix timestamp for when the task was stopped (the task transitioned from the RUNNING state to the STOPPED state).

pub fn set_stopped_at(self, input: Option<Instant>) -> Self

pub fn stopped_reason(self, input: impl Into<String>) -> Self

The reason that the task was stopped.

pub fn set_stopped_reason(self, input: Option<String>) -> Self

pub fn stopping_at(self, input: Instant) -> Self

The Unix timestamp for when the task stops (transitions from the RUNNING state to STOPPED).

pub fn set_stopping_at(self, input: Option<Instant>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn task_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the task.

pub fn set_task_arn(self, input: Option<String>) -> Self

pub fn task_definition_arn(self, input: impl Into<String>) -> Self

The ARN of the task definition that creates the task.

pub fn set_task_definition_arn(self, input: Option<String>) -> Self

pub fn version(self, input: i64) -> Self

The version counter for the task. Every time a task experiences a change that triggers a CloudWatch event, the version counter is incremented. If you are replicating your Amazon ECS task state with CloudWatch Events, you can compare the version of a task reported by the Amazon ECS API actions with the version reported in CloudWatch Events for the task (inside the detail object) to verify that the version in your event stream is current.

pub fn set_version(self, input: Option<i64>) -> Self

pub fn ephemeral_storage(self, input: EphemeralStorage) -> Self

The ephemeral storage settings for the task.

pub fn set_ephemeral_storage(self, input: Option<EphemeralStorage>) -> Self

pub fn build(self) -> Task

Consumes the builder and constructs a Task