Module aws_sdk_ecs::operation

Structs

CreateCapacityProvider
Creates a new capacity provider. Capacity providers are associated with an Amazon ECS

CreateCluster
Creates a new Amazon ECS cluster. By default, your account receives a default

CreateService
Runs and maintains a desired number of tasks from a specified task definition. If the

CreateTaskSet
Create a task set in the specified cluster and service. This is used when a service

DeleteAccountSetting
Disables an account setting for a specified IAM user, IAM role, or the root user for

DeleteAttributes
Deletes one or more custom attributes from an Amazon ECS resource.

DeleteCapacityProvider
Deletes the specified capacity provider.

DeleteCluster
Deletes the specified cluster. The cluster will transition to the

DeleteService
Deletes a specified service within a cluster. You can delete a service if you have no

DeleteTaskSet
Deletes a specified task set within a service. This is used when a service uses the

DeregisterContainerInstance
Deregisters an Amazon ECS container instance from the specified cluster. This instance is

DeregisterTaskDefinition
Deregisters the specified task definition by family and revision. Upon deregistration,

DescribeCapacityProviders
Describes one or more of your capacity providers.

DescribeClusters
Describes one or more of your clusters.

DescribeContainerInstances
Describes Amazon Elastic Container Service container instances. Returns metadata about registered and

DescribeServices
Describes the specified services running in your cluster.

DescribeTaskDefinition
Describes a task definition. You can specify a family and

DescribeTaskSets
Describes the task sets in the specified cluster and service. This is used when a

DescribeTasks
Describes a specified task or tasks.

DiscoverPollEndpoint

ExecuteCommand
Runs a command remotely on a container within a task.

ListAccountSettings
Lists the account settings for a specified principal.

ListAttributes
Lists the attributes for Amazon ECS resources within a specified target type and cluster.

ListClusters
Returns a list of existing clusters.

ListContainerInstances
Returns a list of container instances in a specified cluster. You can filter the

ListServices
Returns a list of services. You can filter the results by cluster, launch type, and

ListTagsForResource
List the tags for an Amazon ECS resource.

ListTaskDefinitionFamilies
Returns a list of task definition families that are registered to your account (which

ListTaskDefinitions
Returns a list of task definitions that are registered to your account. You can filter

ListTasks
Returns a list of tasks. You can filter the results by cluster, task definition

PutAccountSetting
Modifies an account setting. Account settings are set on a per-Region basis.

PutAccountSettingDefault
Modifies an account setting for all IAM users on an account for whom no individual

PutAttributes
Create or update an attribute on an Amazon ECS resource. If the attribute does not exist,

PutClusterCapacityProviders
Modifies the available capacity providers and the default capacity provider strategy

RegisterContainerInstance

RegisterTaskDefinition
Registers a new task definition from the supplied family and

RunTask
Starts a new task using the specified task definition.

StartTask
Starts a new task from the specified task definition on the specified container

StopTask
Stops a running task. Any tags associated with the task will be deleted.

SubmitAttachmentStateChanges

SubmitContainerStateChange

SubmitTaskStateChange

TagResource
Associates the specified tags to a resource with the specified

UntagResource
Deletes specified tags from a resource.

UpdateCapacityProvider
Modifies the parameters for a capacity provider.

UpdateCluster
Updates the cluster.

UpdateClusterSettings
Modifies the settings to use for a cluster.

UpdateContainerAgent
Updates the Amazon ECS container agent on a specified container instance. Updating the

UpdateContainerInstancesState
Modifies the status of an Amazon ECS container instance.

UpdateService

UpdateServicePrimaryTaskSet
Modifies which task set in a service is the primary task set. Any parameters that are

UpdateTaskSet
Modifies a task set. This is used when a service uses the EXTERNAL